### 2025-06-24 28b5d53
Update paper.md
Fix ORCID links for joss review
