### 2025-06-24 #8461 @kyleniemeyer
Hello @justinsgray @SichengHe and @kylebeggs are you available to review this submission to JOSS in the next few weeks? `TASOPT.jl` is a Julia package for conceptual commercial transport aircraft design. Thanks!
