### 2025-06-30 d127407
Better comments and docstrings for new features
