### 2025-07-01 #8521 @editorialbot
[REVIEW]: TASOPT.jl: A Julia package for conceptual commercial transport aircraft design
**Submitting author:** @askprash (<a href="http://orcid.org/0000-0002-2346-0697">Prakash Prashanth</a>)
**Repository:** https://github.com/MIT-LAE/TASOPT.jl
**Branch with paper.md** (empty if default branch): joss
**Version:** v3.0.0
