### 2025-07-03 8e5b6fa
moved Weng to LHS in update_weights
