### 2025-07-03 2cef1dd
different FC example + ipdes for HX check
