+++ example/example_fuel_cell.toml
@@ -75,7 +75,7 @@ name = "TASOPT Model with cryo fuel and HX"
     insulation_material = ["polyurethane32", "polyurethane32", "polyurethane32"] 
     insulation_segment_base_thickness = [0.05, 0.05, 0.05] #m
     
-    cruise_boiloff_rate = 0.4 #%/hour, used if sizes_insulation == true
+    cruise_boiloff_rate = 2.0 #%/hour, used if sizes_insulation == true
     insulation_thicknesses_design_indices = [1,2,3] #insulation segments to size if sizes_insulation == true
 
     #Parameters for double-walled tank, if "vacuum" is in insulation_material
@@ -84,15 +84,15 @@ name = "TASOPT Model with cryo fuel and HX"
 
 [Mission]
     N_missions = 1 # Number of missions to be modeled (first mission is the design mission)
-    payload = ["180 pax"]    # Mission payload; pax can be used as a unit of weight.
+    payload = ["56 pax"]    # Mission payload; pax can be used as a unit of weight.
                              # Weight per pax also specified (215 lbf/pax is std)
-    exit_limit = 189            # Maximum number of passengers that can sit in the cabin
+    exit_limit = 56            # Maximum number of passengers that can sit in the cabin
 
-    max_payload = "189 pax"     # Defines the maximum payload carrying capacity of the aircraft
+    max_payload = "56 pax"     # Defines the maximum payload carrying capacity of the aircraft
                                 # It is the maximum payload that the aircraft can carry (including cargo),
                                 # in an off-design mission
     
-    range = "3000.0 nmi" # Design Range 
+    range = "924.0 nmi" # Design Range 
 
     weight_per_pax = "215.0 lbf" # Specify weight per passenger - 
                             # includes luggage [lbm or lbf or kg or N] 
@@ -125,7 +125,7 @@ name = "TASOPT Model with cryo fuel and HX"
 
 [Mission.Cruise]
     cruise_alt  = "35000.0 ft"
-    cruise_mach = 0.80
+    cruise_mach = 0.60
     cruise_CL   = 0.57
 
 [Mission.Descent]
@@ -136,8 +136,8 @@ name = "TASOPT Model with cryo fuel and HX"
     cabin_pressure_altitude = "8000.0 ft"
 
 [Fuselage.Geometry]
-    radius = "100 in"       # Fuselage radius
-    dRadius = "15 in"      # Downward shift of lower bubbles
+    radius = "1.5 m"       # Fuselage radius
+    dRadius = "0 in"      # Downward shift of lower bubbles
     y_offset = 0.0         # y offset of bubble center
     floor_depth = "5.0 in" # depth of floor beams
     Nwebs = 1              # number of webs (for double bubble designs)
@@ -442,7 +442,7 @@ name = "TASOPT Model with cryo fuel and HX"
     number_of_engines = 2
 
 [Propulsion.DuctedFan]
-    maximum_fan_power = 10e6
+    maximum_fan_power = 3.0e6
     
     rate_of_climb = "2.5 m/s"
 
@@ -461,6 +461,10 @@ name = "TASOPT Model with cryo fuel and HX"
     fan_nozzle_area.descentstart = 1.0
     fan_nozzle_area.descentend   = 1.0
 
+[Propulsion.Nacelles]
+    nacelle_pylon_wetted_area_ratio = 16.0 # = rSnace = wetted area/fan area
+    nacelle_local_velocity_ratio = 1.02 # local/freestream velocity
+
 [Propulsion.FuelCell]
     fuel_cell_type = "HT-PEMFC"
     takeoff_current_density = 10000
[…]
