+++ test/default_structures.jl
@@ -5,7 +5,7 @@ ac_test = load_default_model()
 fuse = ac_test.fuselage
 Weight = TASOPT.structures.Weight
 fuse.n_decks = 1.00000000000000000000 
-fuse.shell.weight = Weight(W = 20950.10880241532868240029 ) 
+fuse.shell.weight = Weight(W = 20950.10880112874656333588 ) 
 fuse.shell.weight.r = [ 16.97078113922184172679 ,0.0,0.0] 
 fuse.window.W = 11269.98000000000138243195 
 fuse.window.r = [ 18.13560000000000016485 ,0.0,0.0] 
@@ -15,12 +15,12 @@ fuse.insulation.r = [ 18.13560000000000016485 ,0.0,0.0]
 fuse.insulation_W_per_area = 22.00000000000000000000 
 fuse.floor.weight.W = 13415.79649570315450546332 
 fuse.floor_W_per_area = 60.00000000000000000000 
-fuse.cone.weight = Weight(W = 5771.54787313527958758641 ) 
+fuse.cone.weight = Weight(W = 5771.54790021659482590621 ) 
 fuse.cone.weight.r = [ 33.37559999999999860165 ,0.0,0.0] 
-fuse.bendingmaterial_h.weight = Weight(W = 14601.36970585678864154033 ) 
-fuse.bendingmaterial_v.weight = Weight(W = 7564.73678495778312935727 ) 
-fuse.weight = 198619.75430119407246820629 
-fuse.moment = 3632749.60595947690308094025 
+fuse.bendingmaterial_h.weight = Weight(W = 14601.36976621137910115067 ) 
+fuse.bendingmaterial_v.weight = Weight(W = 7564.73683054416142113041 ) 
+fuse.weight = 198619.75443292976706288755 
+fuse.moment = 3632749.60894177854061126709 
 fuse.volume = 422.77743761859198912134 
 fuse.weight_frac_stringers = 0.34999999999999997780 
 fuse.weight_frac_frame = 0.25000000000000000000 
@@ -35,25 +35,25 @@ fuse.layout.x_end_cylinder = 29.56559999999999988063
 fuse.layout.x_pressure_shell_fwd = 5.18160000000000042775 
 fuse.layout.x_pressure_shell_aft = 31.08960000000000079012 
 fuse.layout.x_cone_end = 35.66159999999999996589 
-fuse.bendingmaterial_h.weight.r = [ 31.91057796046597161421 ,0.0,0.0] 
-fuse.bendingmaterial_v.weight.r = [ 31.15470082522883643605 ,0.0,0.0] 
+fuse.bendingmaterial_h.weight.r = [ 31.91057797350605440556 ,0.0,0.0] 
+fuse.bendingmaterial_v.weight.r = [ 31.15470084008022055855 ,0.0,0.0] 
 fuse.layout.cross_section.radius = 1.95579999999999998295 
 fuse.layout.cross_section.bubble_lower_downward_shift = 0.38100000000000000533 
 fuse.layout.floor_depth = 0.12700000000000000178 
 fuse.layout.taper_tailcone = 0.29999999999999998890 
 fuse.ratio_young_mod_fuse_bending = 1.00000000000000000000 
-fuse.skin.thickness = 0.00105943102431815751 
-fuse.cone.thickness = 0.00124599710443489844 
+fuse.skin.thickness = 0.00105943102425309606 
+fuse.cone.thickness = 0.00124599711028137842 
 fuse.layout.thickness_webs = 0.00000000000000000000 
 fuse.floor.thickness = 0.00090615659813192320 
-fuse.shell.EIh = 2361630794.38818645477294921875 
[…]
