### 2025-07-08 #8521 @kylebeggs
Is there a reason the package has not been registered in the Julia General Registry so that one can install with `]add TASOPT`?
