### 2025-07-10 662283d
Update README.md install instructions
TASOPT is now registered with the general registry, so can install with just `>pkg add TASOPT`
