### 2025-07-10 5f8d1a2
Update paper.md
