### 2025-07-10 0cac0cd
Update paper.bib
