+++ paper/paper.bib
@@ -1,3 +1,44 @@
+@online{aviary,
+  title = {{{NASA}}'s {{Aviary Takes Flight}}: {{A Public Software}} for {{Aircraft Design}} - {{NASA Technical Reports Server}} ({{NTRS}})},
+  url = {https://ntrs.nasa.gov/citations/20240009217},
+  urldate = {2025-07-10},
+}
+
+@inbook{suave,
+author = {Timothy MacDonald and Matthew Clarke and Emilio M. Botero and Julius M. Vegh and Juan J. Alonso},
+title = {SUAVE: An Open-Source Environment Enabling Multi-Fidelity Vehicle Optimization},
+booktitle = {18th AIAA/ISSMO Multidisciplinary Analysis and Optimization Conference},
+chapter = {},
+pages = {},
+doi = {10.2514/6.2017-4437},
+URL = {https://arc.aiaa.org/doi/abs/10.2514/6.2017-4437},
+eprint = {https://arc.aiaa.org/doi/pdf/10.2514/6.2017-4437}
+}
+
+@inproceedings{openconcept,
+	address = {{C}incinnati,~{OH}},
+	author = {Benjamin J. Brelje and Joaquim R. R. A. Martins},
+	booktitle = {Proceedings of the AIAA/IEEE Electric Aircraft Technologies Symposium},
+	doi = {10.2514/6.2018-4979},
+	month = {July},
+	title = {Development of a Conceptual Design Model for Aircraft Electric Propulsion with Efficient Gradients},
+	year = {2018}
+}
+
+@article{fastOAD,
+    doi = {10.1088/1757-899x/1024/1/012062},
+    url = {https://doi.org/10.1088/1757-899x/1024/1/012062},
+    year = 2021,
+    month = {jan},
+    publisher = {{IOP} Publishing},
+    volume = {1024},
+    number = {1},
+    pages = {012062},
+    author = {Christophe David and Scott Delbecq and Sebastien Defoort and Peter Schmollgruber and Emmanuel Benard and Valerie Pommier-Budinger},
+    title = {From {FAST} to {FAST}-{OAD}: An open source framework for rapid Overall Aircraft Design},
+    journal = {{IOP} Conference Series: Materials Science and Engineering}
+}
+
 @article{prashZIA,
   title = {Near-Zero Environmental Impact Aircraft},
   author = {Prashanth, Prakash and Elmourad, Jad and Grobler, Carla and Isaacs, Stewart and Zahid, Syed Shayan and Abel, James and Falter, Christoph and Fritz, Thibaud and Allroggen, Florian and Sabnis, Jayant S. and Eastham, Sebastian D. and Speth, Raymond L. and Barrett, Steven R. H.},
