### 2025-07-10 cd5f0d0
Update paper.bib
