+++ paper/paper.bib
@@ -2,18 +2,34 @@
   title = {{{NASA}}'s {{Aviary Takes Flight}}: {{A Public Software}} for {{Aircraft Design}} - {{NASA Technical Reports Server}} ({{NTRS}})},
   url = {https://ntrs.nasa.gov/citations/20240009217},
   urldate = {2025-07-10},
+  date = {2024}
 }
 
-@inbook{suave,
-author = {Timothy MacDonald and Matthew Clarke and Emilio M. Botero and Julius M. Vegh and Juan J. Alonso},
-title = {SUAVE: An Open-Source Environment Enabling Multi-Fidelity Vehicle Optimization},
-booktitle = {18th AIAA/ISSMO Multidisciplinary Analysis and Optimization Conference},
-chapter = {},
-pages = {},
-doi = {10.2514/6.2017-4437},
-URL = {https://arc.aiaa.org/doi/abs/10.2514/6.2017-4437},
-eprint = {https://arc.aiaa.org/doi/pdf/10.2514/6.2017-4437}
+@misc{aviaryREPO,
+  title = {{{OpenMDAO}}/{{Aviary}}},
+  year = {2025},
+  month = jul,
+  urldate = {2025-07-10},
+  url = {https://github.com/OpenMDAO/Aviary},
+  abstract = {NASA's aircraft analysis, design, and optimization tool},
+  howpublished = {OpenMDAO},
+  keywords = {aircraft-design,nasa,optimization,trajectory-optimization}
 }
+
+
+@incollection{suave,
+  title = {{{SUAVE}}: {{An Open-Source Environment Enabling Multi-Fidelity Vehicle Optimization}}},
+  shorttitle = {{{SUAVE}}},
+  booktitle = {18th {{AIAA}}/{{ISSMO Multidisciplinary Analysis}} and {{Optimization Conference}}},
+  author = {MacDonald, Timothy and Clarke, Matthew and Botero, Emilio M. and Vegh, Julius M. and Alonso, Juan J.},
+  date = {2017-06-02},
+  series = {{{AIAA AVIATION Forum}}},
+  publisher = {{American Institute of Aeronautics and Astronautics}},
+  doi = {10.2514/6.2017-4437},
+  url = {https://arc.aiaa.org/doi/10.2514/6.2017-4437},
+  urldate = {2025-07-10},
+}
+
 
 @inproceedings{openconcept,
 	address = {{C}incinnati,~{OH}},
