### 2025-07-10 f7b0534
Update paper.md
