### 2025-07-10 fa3de40
Update index.md
update install instructions in docs too
