### 2025-07-17 #8521 @JustinSGray
Im good with the paper as is now.
