### 2025-07-17 #8521 @askprash
Thanks @kylebeggs! It took me embarrassingly long to actually understand how to add type parameters in structs properly. Some of our planned updates and draft PRs are going to try and take advantage of this even more!
