### 2025-07-25 5780509
Change order of reading fuselage data
