### 2025-07-25 4d21930
update order of creating fuselage to set up right parametric types
