### 2025-07-25 4273d82
Proper fix for type stability in fuselage
