### 2025-07-25 178404e
Remove commented line
Left this by accident when I made the previous commits
