### 2025-07-25 5a55ae4
kwdef works too
I was over zealous and removed the kwdef macro but it works just fine.
