+++ src/data_structs/layout.jl
@@ -99,42 +99,37 @@ Contains dimensions, heights, etc. to design a fuselage
 
 $TYPEDFIELDS
 """
-mutable struct FuselageLayout{CS<:AbstractCrossSection} <: AbstractLayout
+@kwdef mutable struct FuselageLayout{CS<:AbstractCrossSection} <: AbstractLayout
     """Cross section definition"""
-    cross_section::CS
+    cross_section::CS = CS()
     """Thickness of webs """
-    thickness_webs::Float64
+    thickness_webs::Float64 = 0.0
     """X position of nose [m]"""
-    x_nose::Float64  
+    x_nose::Float64 = 0.0
     """X position of pressure shell forward [m]"""
-    x_pressure_shell_fwd::Float64 
+    x_pressure_shell_fwd::Float64 = 0.0
     """X position of pressure shell aft [m]"""
-    x_pressure_shell_aft::Float64 
+    x_pressure_shell_aft::Float64 = 0.0
     """X position of cylinder start [m]"""
-    x_start_cylinder::Float64 
+    x_start_cylinder::Float64 = 0.0
     """X position of cylinder end [m]"""
-    x_end_cylinder::Float64 
+    x_end_cylinder::Float64 = 0.0
     """X position of fuselage cone end [m]"""
-    x_cone_end::Float64 
+    x_cone_end::Float64 = 0.0
     """X position of fuselage end [m]"""
-    x_end::Float64 
+    x_end::Float64 = 0.0
     """Tailcone taper (lambdac) [m]"""
-    taper_tailcone::Float64 # lambdac
+    taper_tailcone::Float64 = 0.0 # lambdac
     """Floor depth (depth of floor beams) [m]"""
-    floor_depth::Float64 
+    floor_depth::Float64 = 0.0
     """Nose Radius [m]"""
-    nose_radius::Float64 
+    nose_radius::Float64 = 0.0
     """Tail Radius [m]"""
-    tail_radius::Float64 
+    tail_radius::Float64 = 0.0
     """Taper fuselage to "point" or "edge" """
-    opt_tapers_to::String
+    opt_tapers_to::String = "point"
     """Length of cylindrical portion of cabin that contains payload [m]"""
-    l_cabin_cylinder::Float64
-end
-
-function FuselageLayout{CS}() where {CS<:AbstractCrossSection}
-    return FuselageLayout{CS}(CS(), 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0,
-    "point", 0.0)
+    l_cabin_cylinder::Float64 = 0.0
 end
 
 # Helper function to be able to simplify 
