### 2025-07-25 a8ae828
fixed bug in thermal_params typing
