### 2025-07-25 c2f905b
more robust tank residuals
