+++ src/cryo_tank/tankWthermal.jl
@@ -84,7 +84,15 @@ function tankWthermal(fuse::Fuselage, fuse_tank::fuselage_tank, z::Float64, TSL:
       thickness = sum(t_cond)  # total thickness of insulation
       ΔT = Taw - Tfuel
       
-      fun(x) = residuals_Q(x, p, "Q_unknown") #Create function handle to be zeroed
+      #Function to zero in solver
+      function residual(x) 
+            try
+                  return residuals_Q(x, p, "Q_unknown") #This should be zeroed
+            catch #Return some high residual if it fails
+                  return ones(length(x))*1e3
+            end
+      end
+
       #Initial guess for function
       guess = zeros(length(t_cond) + 2) 
 
@@ -96,7 +104,7 @@ function tankWthermal(fuse::Fuselage, fuse_tank::fuselage_tank, z::Float64, TSL:
             guess[i + 2] = Tfuel + ΔT * sum(t_cond[1:i])/ thickness
       end
       guess[end] = guess[end] - 1.0 #fuselage wall temperature
-      sol = nlsolve(fun, guess, xtol = 1e-7, ftol = 1e-6) #Solve non-linear problem with NLsolve.jl
+      sol = nlsolve(residual, guess, xtol = 1e-7, ftol = 1e-6) #Solve non-linear problem with NLsolve.jl
       
       Q = qfac * sol.zero[1]    # Heat rate from ambient to cryo fuel, including extra heat leak from valves etc as in eq 3.20 by Verstraete
       
