### 2025-08-12 #8521 @askprash
@kyleniemeyer thanks for checking in. I've updated all the requested changes based on the reviewer feedback with the exception of the broken docs badge link in the README. Once the paper is accepted I'll update the rc tag which should fix the `stable` vs `dev` docs badges. I'd like to link to the JOSS paper in the release notes, if that makes sense. Let me know if I need to take any action. Thanks!
