### 2025-08-13 033e70d
Updates JOSS paper
Fixes some in-text reference formatting
