+++ paper/paper.md
@@ -51,17 +51,17 @@ bibliography: paper.bib
 
 # Summary
 
-Aircraft design is a multidisciplinary field that spans aerodynamics, structures, thermodynamics and propulsion. Prior approaches to this problem have consisted of historical correlations, empirical drag and simplified propulsion system models [@roskam], [@Raymer], [@Torenbeek]. 
+Aircraft design is a multidisciplinary field that spans aerodynamics, structures, thermodynamics and propulsion. Prior approaches to this problem have consisted of historical correlations, empirical drag and simplified propulsion system models [@roskam; @Raymer; @Torenbeek]. 
 
-TASOPT, originally written in FORTRAN and developed by Mark Drela [@Drela_2011], [@Greitzer2010], took the approach of basing the aerodynamic, propulsion, and weight estimates on low-order models built on first-principles structural, aerodynamic, thermodynamic theory to the extent possible. This approach gives confidence that optimized aircraft designs generated by the program are not spurious artifacts of inappropriately extrapolated empirical data. 
+TASOPT, originally written in FORTRAN and developed by Mark Drela [@Drela_2011; @Greitzer2010], took the approach of basing the aerodynamic, propulsion, and weight estimates on low-order models built on first-principles structural, aerodynamic, thermodynamic theory to the extent possible. This approach gives confidence that optimized aircraft designs generated by the program are not spurious artifacts of inappropriately extrapolated empirical data. 
 
-This package, `TASOPT.jl`, is a modern, Julia based implementation of the same principles, written with the goal of extensibility in mind. The current package leverages open source Julia packages where useful to provide an easy-to-read and maintain software. `TASOPT.jl` also introduces additional functionality such as capability to model some forms of hydrogen fuel aircraft. In addition, the software allows users to connect it with other aircraft propulsion system tools (such as NPSS) and aircraft environmental impact assessment tools (such as Aviation Emissions Inventory Codes [@simone2013]).
+This package, `TASOPT.jl`, is a modern, Julia based implementation of the same principles, written with the goal of extensibility in mind. The current package leverages open source Julia packages where useful to provide an easy-to-read and maintain software. `TASOPT.jl` also introduces additional functionality such as capability to model some forms of hydrogen fuel aircraft. In addition, the software allows users to connect it with other aircraft propulsion system tools (such as NPSS) and aircraft environmental impact assessment tools (such as Aviation Emissions Inventory Codes) [@simone2013].
 
 # Statement of need
 
 A fundamental challenge faced by the aviation sector is to minimize its impact on the environment. There are numerous research groups around the world working on various system-level solutions to this challenge. However, these research groups are often forced to resort to using simplified black-box descriptions of aircraft performance such as BADA that are limited to modeling aircraft that are currently present in the fleet. `TASOPT.jl` provides an alternative physics-based model that allows researchers to better represent current and future aircraft performance in their system level models.
 
-Most modern aircraft design tools generally fall into a few categories: high fidelity tools (ADflow in openMDAO, SU2, Cart3D) that require extensive compute resources, intermediate fidelity tools (e.g., NASA's Aviary tool set [@aviary], [@aviaryREPO], Stanford's SUAVE [@suave], Michigan's Open-Concept [@openconcept], Onera's FAST-OAD [@fastOAD]) or simplified GUI based tools (ADS, openVSP, RDSwin) that rely on approximations and historical empiricism.  `TASOPT.jl` is a Julia package that provides an easy-to-use interface for conceptual aircraft multidisciplinary design and optimization (MDAO) in a high-level language with minimal compromise on computational speed. 
[…]
