### 2025-08-13 6071822
Updates JOSS bib
Fixes some citations for JOSS paper
