+++ paper/paper.bib
@@ -17,17 +17,15 @@
 }
 
 
-@incollection{suave,
+@inproceedings{suave,
   title = {{{SUAVE}}: {{An Open-Source Environment Enabling Multi-Fidelity Vehicle Optimization}}},
-  shorttitle = {{{SUAVE}}},
   booktitle = {18th {{AIAA}}/{{ISSMO Multidisciplinary Analysis}} and {{Optimization Conference}}},
   author = {MacDonald, Timothy and Clarke, Matthew and Botero, Emilio M. and Vegh, Julius M. and Alonso, Juan J.},
-  date = {2017-06-02},
+  year = {2017},
+  month = jun,
   series = {{{AIAA AVIATION Forum}}},
   publisher = {{American Institute of Aeronautics and Astronautics}},
   doi = {10.2514/6.2017-4437},
-  url = {https://arc.aiaa.org/doi/10.2514/6.2017-4437},
-  urldate = {2025-07-10},
 }
 
 
@@ -60,6 +58,8 @@
   author = {Prashanth, Prakash and Elmourad, Jad and Grobler, Carla and Isaacs, Stewart and Zahid, Syed Shayan and Abel, James and Falter, Christoph and Fritz, Thibaud and Allroggen, Florian and Sabnis, Jayant S. and Eastham, Sebastian D. and Speth, Raymond L. and Barrett, Steven R. H.},
   year = {2024},
   month = jul,
+  volume = {8},
+  pages = {4772--4782},
   journal = {Sustainable Energy \& Fuels},
   publisher = {The Royal Society of Chemistry},
   issn = {2398-4902},
@@ -99,20 +99,30 @@
   urldate = {2024-11-24},
 }
 
-@article{Greitzer2010, 
- title={N + 3 Aircraft Concept Designs and Trade Studies , Final Report Volume 1},
- volume={1}, 
- note={Citation Key: Greitzer2010a}, number={March},
- author={Greitzer, E. M. and Bonnefoy, P. A. and De La Rosa Blanco, E. and Dorbian, C. S. and Drela, M. and Hall, D. K. and Hansman, R. J. and Hileman, J. I. and Liebeck, R. H. and Lovegren, J. and Mody, P. and Pertuze, J. a. and Sato, S. and Spakovszky, Z. S. and Tan, C. S. and Hollman, J. S.},
- year={2010}, pages={1–187} }
+@techreport{Greitzer2010, 
+ title={N + 3 Aircraft Concept Designs and Trade Studies, Final Report Volume 1},
+ institution={NASA},
+ author={Greitzer, E. M. and Bonnefoy, P. A. and De La Rosa Blanco, E. and Dorbian, C. S. and Drela, M. and 
+Hall, D. K. and Hansman, R. J. and Hileman, J. I. and Liebeck, R. H. and Lovegren, J. and Mody, P. and 
+Pertuze, J. a. and Sato, S. and Spakovszky, Z. S. and Tan, C. S. and Hollman, J. S. and Duda, J. E. and 
+Fitzgerald, N. and Houghton, J. and Kerrebrock, J. L. and Kiwada, G. F. and Kordonowy, D. and Parrish, J. C. and
+Tylko, J. and Wen, E. A. and Lord, W. K.},
+ year={2010},
+ pages={1--187},
+ number={NASA/CR—2010-216794/VOL1},
+ url={https://ntrs.nasa.gov/citations/20100042401}
+}
 
- @article{Drela_2011, 
- title={Development of the D8 Transport Configuration}, 
+ @inproceedings{Drela_2011, 
+ title={Development of the {D8} Transport Configuration}, 
  ISBN={978-1-62410-145-8},
- DOI={10.2514/6.2011-3970}, 
[…]
