### 2025-08-13 #8521 @kyleniemeyer
@askprash I've made some fixes to the JOSS paper in https://github.com/MIT-LAE/TASOPT.jl/pull/185, please review and merge.
