### 2025-08-13 #8521 @kyleniemeyer
@askprash forgot to mention: in the paper, I noticed a few software packages mentioned that were not cited ("ADS, openVSP, RDSwin"), though other works in the same sentence *were* cited. Could you add citations for these?
