### 2025-08-13 0a0924c
Update paper.bib
Add missed bib items
