+++ paper/paper.bib
@@ -1,3 +1,17 @@
+@online{OpenVSP,
+  title = {{{OpenVSP}}/{{OpenVSP}}},
+  url = {https://github.com/OpenVSP/OpenVSP},
+  urldate = {2025-08-13},
+  date = {2025}
+}
+
+@online{ADS,
+  title = {{{ADS}} - {{Aircraft Design Software}}, {{OAD}} - {{Optimal Aircraft Design}}},
+  url = {https://www.pca2000.com/},
+  urldate = {2025-08-13},
+}
+
+
 @online{aviary,
   title = {{{NASA}}'s {{Aviary Takes Flight}}: {{A Public Software}} for {{Aircraft Design}} - {{NASA Technical Reports Server}} ({{NTRS}})},
   url = {https://ntrs.nasa.gov/citations/20240009217},
@@ -138,7 +152,24 @@ Tylko, J. and Wen, E. A. and Lord, W. K.},
   }
 
 
- @misc{RDS-win, title={Aircraft Design Software}, url={http://www.aircraftdesign.com/rds.shtml} }
+ @online{RDS-win, 
+    title={Aircraft Design Software},
+    author = {Daniel Raymer},
+    url={http://www.aircraftdesign.com/rds.shtml},
+}
+
+@inproceedings{raymer2011,
+  title = {Conceptual {{Design Modeling}} in the {{RDS-Professional Aircraft Design Software}}},
+  booktitle = {49th {{AIAA Aerospace Sciences Meeting}} Including the {{New Horizons Forum}} and {{Aerospace Exposition}}},
+  author = {Raymer, Dan},
+  date = {2011-01-04},
+  series = {Aerospace {{Sciences Meetings}}},
+  publisher = {{American Institute of Aeronautics and Astronautics}},
+  doi = {10.2514/6.2011-161},
+  url = {https://arc.aiaa.org/doi/10.2514/6.2011-161},
+  urldate = {2025-08-13}
+}
+
  
  @book{Raymer, address={Reston, VA}, 
  series={AIAA education series}, 
