### 2025-08-13 6c5e90b
Update paper.md
Add missed citations
