+++ paper/paper.md
@@ -61,7 +61,7 @@ This package, `TASOPT.jl`, is a modern, Julia based implementation of the same p
 
 A fundamental challenge faced by the aviation sector is to minimize its impact on the environment. There are numerous research groups around the world working on various system-level solutions to this challenge. However, these research groups are often forced to resort to using simplified black-box descriptions of aircraft performance such as BADA that are limited to modeling aircraft that are currently present in the fleet. `TASOPT.jl` provides an alternative physics-based model that allows researchers to better represent current and future aircraft performance in their system level models.
 
-Most modern aircraft design tools generally fall into a few categories: high fidelity tools (ADflow in openMDAO, SU2, Cart3D) that require extensive compute resources, intermediate fidelity tools (e.g., NASA's Aviary tool set [@aviary; @aviaryREPO], Stanford's SUAVE [@suave], Michigan's Open-Concept [@openconcept], Onera's FAST-OAD [@fastOAD]) or simplified GUI based tools (ADS, openVSP, RDSwin) that rely on approximations and historical empiricism.  `TASOPT.jl` is a Julia package that provides an easy-to-use interface for conceptual aircraft multidisciplinary design and optimization (MDAO) in a high-level language with minimal compromise on computational speed. 
+Most modern aircraft design tools generally fall into a few categories: high fidelity tools (ADflow in openMDAO, SU2, Cart3D) that require extensive compute resources, intermediate fidelity tools (e.g., NASA's Aviary tool set [@aviary; @aviaryREPO], Stanford's SUAVE [@suave], Michigan's Open-Concept [@openconcept], Onera's FAST-OAD [@fastOAD]) or simplified GUI based tools (ADS [@ADS], openVSP [@OpenVSP], RDSwin [@RDS-win; @raymer2011]) that rely on approximations and historical empiricism.  `TASOPT.jl` is a Julia package that provides an easy-to-use interface for conceptual aircraft multidisciplinary design and optimization (MDAO) in a high-level language with minimal compromise on computational speed. 
 
 `TASOPT.jl` is designed to be used by researchers in aircraft and propulsion system design as well as educators and students (at the upper undergraduate to graduate level) in courses to understand fundamental aircraft design and optimization tradeoffs. Being written in a modern high-level language (Julia) reduces the barrier to entry for young researchers and students who wish to explore intermediate fidelity aircraft MDAO.
 
