### 2025-08-21 6ac6c40
Update paper.md
Add Adi paper citation
