+++ paper/paper.md
@@ -65,7 +65,7 @@ Most modern aircraft design tools generally fall into a few categories: high fid
 
 `TASOPT.jl` is designed to be used by researchers in aircraft and propulsion system design as well as educators and students (at the upper undergraduate to graduate level) in courses to understand fundamental aircraft design and optimization tradeoffs. Being written in a modern high-level language (Julia) reduces the barrier to entry for young researchers and students who wish to explore intermediate fidelity aircraft MDAO.
 
-`TASOPT.jl` is intended to be easily integrated into existing research workflows due to the ease of integrating Julia code into existing Python code or driving other executable codes such as NPSS engine design. `TASOPT.jl` has been used by several of the authors in past and ongoing projects that studied impact of hydrogen tank design assumptions on aircraft performance [@nicoH2], hydrogen burning turbofans with heat-exchangers [@nicoTank], the infulence of turbofan design parameters on aviation environmental impact [@leeTurbofan], a near-zero impact aircraft system [@prashZIA], and forthcoming theses by Aditeya Shukla and Jonas Gonzalez. Planned future work includes several performance updates, further modularization, integration with `ModelingToolkit.jl`, and integrating noise calculations using `pyNA` [@pyna] to estimate certification landing and take-off noise for aircraft. 
+`TASOPT.jl` is intended to be easily integrated into existing research workflows due to the ease of integrating Julia code into existing Python code or driving other executable codes such as NPSS engine design. `TASOPT.jl` has been used by several of the authors in past and ongoing projects that studied impact of hydrogen tank design assumptions on aircraft performance [@nicoH2], hydrogen burning turbofans with heat-exchangers [@nicoTank], the infulence of turbofan design parameters on aviation environmental impact [@leeTurbofan], a near-zero impact aircraft system [@prashZIA], the climate and air quality benefits of aircraft technology improvements [@adi2025], and forthcoming theses by Jonas Gonzalez. Planned future work includes several performance updates, further modularization, integration with `ModelingToolkit.jl`, and integrating noise calculations using `pyNA` [@pyna] to estimate certification landing and take-off noise for aircraft. 
 
 # Acknowledgements
 
