+++ paper/paper.bib
@@ -42,6 +42,14 @@
   doi = {10.2514/6.2017-4437},
 }
 
+@inproceedings{adi2025,
+  title = {Impact of {{Efficiency-Driven Aircraft Technology Improvements}} on {{Climate}} and {{Air Quality}}},
+  booktitle = {{{AIAA AVIATION FORUM AND ASCEND}} 2025},
+  author = {Shukla, Aditeya and Prashanth, Prakash and Speth, Raymond L.},
+  publisher = {{American Institute of Aeronautics and Astronautics}},
+  doi = {10.2514/6.2025-3502},
+  url = {https://arc.aiaa.org/doi/abs/10.2514/6.2025-3502},
+ }
 
 @inproceedings{openconcept,
 	address = {{C}incinnati,~{OH}},
