### 2025-08-21 7754266
Update paper.bib
