+++ paper/paper.bib
@@ -48,6 +48,7 @@
   author = {Shukla, Aditeya and Prashanth, Prakash and Speth, Raymond L.},
   publisher = {{American Institute of Aeronautics and Astronautics}},
   doi = {10.2514/6.2025-3502},
+  date = {2025}
   url = {https://arc.aiaa.org/doi/abs/10.2514/6.2025-3502},
  }
 
@@ -64,8 +65,7 @@
 @article{fastOAD,
     doi = {10.1088/1757-899x/1024/1/012062},
     url = {https://doi.org/10.1088/1757-899x/1024/1/012062},
-    year = 2021,
-    month = {jan},
+	date = {2021},
     publisher = {{IOP} Publishing},
     volume = {1024},
     number = {1},
