### 2025-08-21 5c880ad
Update paper.bib
fix comma
