+++ paper/paper.bib
@@ -48,7 +48,7 @@
   author = {Shukla, Aditeya and Prashanth, Prakash and Speth, Raymond L.},
   publisher = {{American Institute of Aeronautics and Astronautics}},
   doi = {10.2514/6.2025-3502},
-  date = {2025}
+  date = {2025},
   url = {https://arc.aiaa.org/doi/abs/10.2514/6.2025-3502},
  }
 
