+++ v3.0.0/assets/documenter.js
@@ -1,0 +1,1199 @@
+// Generated by Documenter.jl
+requirejs.config({
+  paths: {
+    'highlight-julia': 'https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.8.0/languages/julia.min',
+    'headroom': 'https://cdnjs.cloudflare.com/ajax/libs/headroom/0.12.0/headroom.min',
+    'jqueryui': 'https://cdnjs.cloudflare.com/ajax/libs/jqueryui/1.13.2/jquery-ui.min',
+    'katex-auto-render': 'https://cdnjs.cloudflare.com/ajax/libs/KaTeX/0.16.8/contrib/auto-render.min',
+    'jquery': 'https://cdnjs.cloudflare.com/ajax/libs/jquery/3.7.0/jquery.min',
+    'headroom-jquery': 'https://cdnjs.cloudflare.com/ajax/libs/headroom/0.12.0/jQuery.headroom.min',
+    'katex': 'https://cdnjs.cloudflare.com/ajax/libs/KaTeX/0.16.8/katex.min',
+    'highlight': 'https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.8.0/highlight.min',
+    'highlight-julia-repl': 'https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.8.0/languages/julia-repl.min',
+  },
+  shim: {
+  "highlight-julia": {
+    "deps": [
+      "highlight"
+    ]
+  },
+  "katex-auto-render": {
+    "deps": [
+      "katex"
+    ]
+  },
+  "headroom-jquery": {
+    "deps": [
+      "jquery",
+      "headroom"
+    ]
+  },
+  "highlight-julia-repl": {
+    "deps": [
+      "highlight"
+    ]
+  }
+}
+});
+////////////////////////////////////////////////////////////////////////////////
+require(['jquery', 'katex', 'katex-auto-render'], function($, katex, renderMathInElement) {
+$(document).ready(function() {
+  renderMathInElement(
+    document.body,
+    {
+  "delimiters": [
+    {
+      "left": "$",
+      "right": "$",
+      "display": false
+    },
+    {
+      "left": "$$",
+      "right": "$$",
+      "display": true
+    },
+    {
+      "left": "\\[",
+      "right": "\\]",
+      "display": true
+    }
+  ],
+  "macros": {
+    "\\Pstd": "P_{\\mathrm{std}}",
+    "\\Ru": "R_{\\mathrm{univ.}}",
+    "\\genfuel": "{\\mathrm{C}_{x_{\\mathrm{C}}}\\mathrm{H}_{x_{\\mathrm{H}}}\\mathrm{O}_{x_{\\mathrm{O}}}\\mathrm{N}_{x_{\\mathrm{N}}}}",
+    "\\Tstd": "T_{\\mathrm{std}}"
+  }
+}
+
+  );
+})
+
+})
+////////////////////////////////////////////////////////////////////////////////
+require(['jquery', 'highlight', 'highlight-julia', 'highlight-julia-repl'], function($) {
+$(document).ready(function() {
+    hljs.highlightAll();
+})
+
+})
+////////////////////////////////////////////////////////////////////////////////
+require(['jquery'], function($) {
+
+let timer = 0;
+var isExpanded = true;
+
+$(document).on(
+  "click",
+  ".docstring .docstring-article-toggle-button",
+  function () {
+    let articleToggleTitle = "Expand docstring";
+    const parent = $(this).parent();
+
+    debounce(() => {
+      if (parent.siblings("section").is(":visible")) {
+        parent
+          .find("a.docstring-article-toggle-button")
+          .removeClass("fa-chevron-down")
+          .addClass("fa-chevron-right");
+      } else {
+        parent
[…]
