+++ versions.js
@@ -1,5 +1,7 @@
 var DOC_VERSIONS = [
+  "stable",
+  "v3.0",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "dev";
-var DOCUMENTER_STABLE = "dev";
+var DOCUMENTER_NEWEST = "v3.0.0";
+var DOCUMENTER_STABLE = "stable";
