### 2025-08-26 #8521 @askprash
Hi @kyleniemeyer, sorry for the delay.
- [x] v3.0.0 is tagged and released
- [x] Tagged version is archived on zenodo doi: [10.5281/zenodo.16950138](https://doi.org/10.5281/zenodo.16950138)
