### 2025-09-02 5f03dbf
Update paper.bib
Fixed author name inconsistency in biblio
