+++ paper/paper.bib
@@ -162,14 +162,14 @@ Tylko, J. and Wen, E. A. and Lord, W. K.},
 
  @online{RDS-win, 
     title={Aircraft Design Software},
-    author = {Daniel Raymer},
+    author={Raymer, Daniel P.}, 
     url={http://www.aircraftdesign.com/rds.shtml},
 }
 
 @inproceedings{raymer2011,
   title = {Conceptual {{Design Modeling}} in the {{RDS-Professional Aircraft Design Software}}},
   booktitle = {49th {{AIAA Aerospace Sciences Meeting}} Including the {{New Horizons Forum}} and {{Aerospace Exposition}}},
-  author = {Raymer, Dan},
+  author={Raymer, Daniel P.}, 
   date = {2011-01-04},
   series = {Aerospace {{Sciences Meetings}}},
   publisher = {{American Institute of Aeronautics and Astronautics}},
