### 2025-09-03 630a11d
Add JOSS badge to README.md
