+++ README.md
@@ -1,5 +1,6 @@
 # TASOPT.jl
-[![](https://img.shields.io/badge/docs-latest-blue.svg)](https://mit-lae.github.io/TASOPT.jl/dev/) [![](https://img.shields.io/badge/docs-stable-blue.svg)](https://mit-lae.github.io/TASOPT.jl/stable/) [![CI](https://github.com/MIT-LAE/TASOPT.jl/actions/workflows/CI.yml/badge.svg?branch=main&job=test&version=lts)](https://github.com/MIT-LAE/TASOPT.jl/actions/workflows/CI.yml) [![codecov](https://codecov.io/github/MIT-LAE/TASOPT.jl/graph/badge.svg?token=J1FNXGO3SD)](https://codecov.io/github/MIT-LAE/TASOPT.jl) 
+[![](https://img.shields.io/badge/docs-latest-blue.svg)](https://mit-lae.github.io/TASOPT.jl/dev/) [![](https://img.shields.io/badge/docs-stable-blue.svg)](https://mit-lae.github.io/TASOPT.jl/stable/) [![CI](https://github.com/MIT-LAE/TASOPT.jl/actions/workflows/CI.yml/badge.svg?branch=main&job=test&version=lts)](https://github.com/MIT-LAE/TASOPT.jl/actions/workflows/CI.yml) [![codecov](https://codecov.io/github/MIT-LAE/TASOPT.jl/graph/badge.svg?token=J1FNXGO3SD)](https://codecov.io/github/MIT-LAE/TASOPT.jl) [![DOI](https://joss.theoj.org/papers/10.21105/joss.08521/status.svg)](https://doi.org/10.21105/joss.08521)
+
 
 Transport Aircraft and the Environment System OPTimization (TASOPT) implemented in Julia. Originally based on Mark Drela's [FORTRAN code](https://web.mit.edu/drela/Public/web/tasopt/) of the same name.
 
