### 2025-09-03 13e7caa
Create CITATION.cff
