### 2025-09-05 06590a9
updates for compatibility with new models
