+++ src/sizing/size_aircraft.jl
@@ -886,7 +886,7 @@ function _size_aircraft!(ac; itermax=35,
         @warn "Some engine points did not converge"
     end
     #Warn user if HX effectiveness is overwritten
-    check_HX_overwriting(engine.heat_exchangers) 
+    check_HX_overwriting(eng.heat_exchangers) 
 end
 
 #TODO: update_WMTO! and update_weights! docstrings need full description
