+++ src/engine/engine.jl
@@ -42,7 +42,6 @@ include("turbofan/tfweight.jl")
 include("turbofan/tfwrap.jl")
 include("turbofan/tfweightwrap.jl")
 include("hxfun.jl")
-include("PEMfuelcell.jl")
 include("simple_engine/constant_TSFC_engine.jl")
 include("simple_engine/fractional_engine_weight.jl")
 include(joinpath(__TASOPTroot__,"data_structs/engine.jl"))
