### 2025-09-05 7e0ee2e
added docs and simplified constant TSFC engine
