### 2025-09-18 7ad021a
added `takeoff!()` call to `fly_mission!()`. tweaks and clarifications in aero docs
