### 2025-09-18 7f42e1e
added `takeoff!()` call to `fly_mission!()`. tweaks and clarifications in aero docs
