+++ src/sizing/size_aircraft.jl
@@ -522,8 +522,8 @@ function _size_aircraft!(ac; itermax=35,
         # Calculate wing engine weight
         if compare_strings(options.opt_engine_location,"wing")
             if compare_strings(options.opt_prop_sys_arch,"te")
-                @error "Support for turboelectric architectures is not currently supported. Their reintroduction with `struct`s is on the roadmap."
-            elseif compare_strings(options.opt_prop_sys_arch,"tf") || compare_strings(options.opt_prop_sys_arch,"constant_tsfc")
+                @error "Turboelectric architectures are not currently supported. Their reintroduction with `struct`s is on the roadmap."
+            elseif compare_strings(options.opt_prop_sys_arch,"tf")
                 Weng1 = parg[igWeng] / parg[igneng]
             end
         else
