### 2025-09-19 c00f90c
added first functional draft of `aeroperf_sweep` and `DragPolar` to present lift-drag curves
