### 2025-09-19 f69c788
added first functional draft of `aeroperf_sweep` and `DragPolar` to present lift-drag curves
