### 2025-09-19 d61e47e
tweaks to have `takeoff!` analysis during `fly_mission!()`
