### 2025-09-22 c49624e
updates following PR review
