### 2025-09-23 7d91e56
more robust engine cases in IO
