### 2025-09-23 24dccf5
removed turboelectric from options
