### 2025-09-23 6759109
added airfoil drag polar plot option, drag polar docstring
