### 2025-09-23 902d93b
added airfoil drag polar plot option, drag polar docstring
