### 2025-09-24 0a257f1
- added plot_airf alias and docstring - addressed export issue with balance_aircraft!(); exported plot_airf and DragPolar - doc tweaks
