### 2025-09-24 3de3c3c
- added plot_airf alias and docstring - addressed export issue with balance_aircraft!(); exported plot_airf and DragPolar - doc tweaks
