+++ src/aero/drag.jl
@@ -18,8 +18,8 @@ where:
 - ``C_{D,fuse}`` (`CDfuse`) is the fuselage profile drage computed by solving a boundary layer integral equation,
 - ``C_{D,wing}`` (`CDwing`) is the wing profile drag (viscous + pressure) computed using airfoil data obtained from CFD,
 - ``C_{D,over}`` (`CDover`) is the fuselage added CD due to lift carryover,
-- ``C_{D,htail}`` (`CDhtail`) is the horizontal tail profile drag computed in a similar manner with `CDwing`,
-- ``C_{D,vtail}`` (`CDvtail`) is the vertical tail profile drag computed in a similar manner with `CDwing`,
+- ``C_{D,htail}`` (`CDhtail`) is the horizontal tail profile drag computed from assumed drag coefficients,
+- ``C_{D,vtail}`` (`CDvtail`) is the vertical tail profile drag computed from assumed drag coefficients,
 - ``C_{D,strut}`` (`CDstrut`) is the strut profile drag, 
 - ``C_{D,nace}`` (`CDnace`) is the nacelle profile drag,
 - ``\\Delta C_{D,BLI,f}`` (`dCDBLIf`) is related to the boundary layer ingestion on the fuselage,
