+++ src/aero/aeroperf_sweep.jl
@@ -1,4 +1,3 @@
-# import ....TASOPT: balance_aircraft!
 using Printf
 
 """
