### 2025-09-24 d0320d3
added aero assumptions and plot summary docs page
