### 2025-09-24 4b58f5a
added aero assumptions and plot summary docs page
