+++ docs/src/aero/moment.md
@@ -1,5 +1,6 @@
 # Pitching moment
-The pitching moments of lifting surfaces are computed by integration of the wing loading with reference to a prescribable wing axis.
+
+The pitching moments of lifting surfaces are computed by integration of the wing loading with reference to a prescribable wing axis. Aircraft stability and balance considerations are detailed [here](@ref stability).
 
 ```@eval
 using Markdown
