### 2025-09-24 ae876e5
added unit tests, tweaks to return statements
