### 2025-09-24 e0128a6
added unit tests, tweaks to return statements
