### 2025-09-27 ed93e31
- range instead of vector
