### 2025-09-28 f6bfc0a
minor clarifications in aerosummary page adding Mach to docs for aeroperf_sweep and dragpolar
