+++ src/IO/output_plots.jl
@@ -1138,6 +1138,7 @@ It produces two side-by-side plots:
     **Inputs:**
     - `ac`: Aircraft model object.
     - `CL_range`: Range of lift coefficients to sweep (default: `0.2:0.05:0.8`).
+    - `Mach`: Mach number at which points are evaluated. If none specified, defaults to value at specified ip, imission (optl.).
     - `show_drag_components::Bool`: If `true`, overlays induced drag and component drag contributions (`CDi`, `CDwing`, `CDfuse`, `CDhtail`, `CDvtail`, `CDother`) on the CL–CD plot (default: `false`).
     - `show_airfoil_data::Bool`: If `true`, overlays airfoil section performance (`clpss`, `cdss`) on both plots (default: `false`).
     - `title::Union{String,Nothing}`: Custom plot title. If `nothing`, uses a default title with aircraft name (default: `nothing`).
