### 2025-09-28 e0d5bc7
minor clarifications in aerosummary page adding Mach to docs for aeroperf_sweep and dragpolar
