+++ docs/src/aero/aerosummary.md
@@ -2,7 +2,6 @@
 
 ## Core assumptions
 
-
 !!! details "📖 In Drela's Words - aerodynamic summary"
 
     The wing airfoil performance is represented by a parameterized transonic airfoil family spanning a range of thicknesses, whose performance is determined by 2D viscous/inviscid CFD calculation for a range of lift coefficients and Mach numbers. Together with suitable sweep corrections, this gives reliable profile+wave drag of the wing in cruise and high climb and high descent. The fuselage drag is likewise obtained from compressible viscous/inviscid CFD, suitably simplified with axisymmetric-based approximations. A side benefit is that detailed knowledge of the fuselage boundary layers makes it possibly for TASOPT to reliably predict the benefits of boundary layer ingestion in fuselage-mounted engines.
@@ -13,11 +12,12 @@ The following table summarizes the modelling assumptions for the aerodynamic con
 
 | Component   | Lift                                     | Drag                                                        | Pitching Moment                                                                 |
 |:------------|:-----------------------------------------|:------------------------------------------------------------|:--------------------------------------------------------------------------------|
-| Wing        | Set by geometry and loading assumptions  | Linked to $C_L$ via 2D CFD and discretization of a finite swept wing | Linked to $C_L$ via 2D CFD and discretization of a finite swept wing. Includes airfoil pitching moment and lift contribution. |
-| Fuselage    | -                                        | Profile drag via axisymmetric IBLT                          | Lift contribution from slender body theory                                      |
+| Wing        | Set by geometry and loading assumptions  | Linked to $C_L$ via 2D CFD and discretization of a finite swept wing† | Integrates discretized lift distribution and assumed airfoil pitching moments |
+| Fuselage    | -                                        | Profile drag via axisymmetric IBLT                          | Lift contribution linearized with prescribed $CMV_{f1}$ per slender body theory                                      |
 | Horz. Stab. | Set by stability requirements            | Finite swept wing for fixed $c_{d,f}$, $c_{d,p}$                  | Linked to $C_{L,h}$ via 2D CFD and discretization of a finite swept wing          |
 | Vert. Stab. | -                                        | Finite swept wing for fixed $c_{d,f}$, $c_{d,p}$                  | -                                                                               |
 | Nacelle     | -                                        | Wetted area with overspeed corrections                      | -                                                                               |
+† Can also use lower-fidelity Reynolds scaling for a finite swept wing with prescribed airfoil quantities, $c_{d,f}$, $c_{d,p}$.
 
 ---
 ## Aircraft drag polars
@@ -29,7 +29,7 @@ The overall aerodynamic performance of the trimmed aircraft can be generated by 
 ---
 ## Airfoil database and summary
 
-The default airfoil section performance database can be found at `./aero/airfoil_data/C.air` and is stored in an [`airfoil`](@ref aerodynamics.airfoil) struct within the `aircraft` struct at `TASOPT.aircraft.wing.airsection`. [`plot_airf()`](@ref) generates a summary plot of the airfoil performance:
+The default airfoil section performance database can be found at `./aero/airfoil_data/C.air` and is stored in an [`airfoil`](@ref aerodynamics.airfoil) struct within the `aircraft` struct at `TASOPT.aircraft.wing.airsection`. A summary plot of the airfoil performance can be generated by [`plot_airf()`](@ref):
 
 ![Airfoil lift, drag, and pitching moment performance.](../assets/plot_airf_output.svg)
 
