### 2025-09-29 3a86cd8
fixed inputs for fuel cells
