+++ test/unit_test_io.jl
@@ -88,7 +88,7 @@
     @test size(csv2,1) == 1 #1 row with addl indices (all)
 
     @test length(csv1[1]) == 763 # = entries w/ full ac `struct` and in default_output_indices
-    @test length(csv2[1]) == 1133 # = entries w/ full ac `struct` and all output_indices
+    @test length(csv2[1]) == 1131 # = entries w/ full ac `struct` and all output_indices
 
     #test the nested vectors within par arrays
     #a: row 1 in both csvs matches the design cruise point/mission 
