+++ src/IO/read_input.jl
@@ -820,13 +820,13 @@ readstruct(x) = read_input(x, structures, dstructures)
 # ---------------------------------
 # Propulsion systems
 # ---------------------------------
-propsys = read_input("prop_sys_arch", options, doptions)
+props = read_input("prop_sys_arch", options, doptions)
 prop = read_input("Propulsion", data, default)
 dprop = default["Propulsion"]
 readprop(x) = read_input(x, prop, dprop)
 parg[igneng] = readprop("number_of_engines")
 
-if compare_strings(propsys,"tf")
+if compare_strings(props,"tf")
         parg[igTmetal] = Temp.(readprop("T_max_metal"))
         parg[igfTt4CL1] = readprop("Tt4_frac_bottom_of_climb")
         parg[igfTt4CLn] = readprop("Tt4_frac_top_of_climb")
@@ -1014,7 +1014,7 @@ if compare_strings(propsys,"tf")
         pare[ieA7fac, iptest, :] .= A7static
         pare[ieA5fac, iptest, :] .= A5static
 
-elseif lowercase(propsys) == "constant_tsfc" #For constant TSFC model
+elseif lowercase(props) == "constant_tsfc" #For constant TSFC model
     ROCdes = readprop("rate_of_climb")
     if ROCdes isa AbstractVector
         para[iaROCdes,ipclimb1:ipclimbn,:] .= [Speed(x) for x in ROCdes]
@@ -1025,57 +1025,7 @@ elseif lowercase(propsys) == "constant_tsfc" #For constant TSFC model
     pare[ieTSFC,ipcruise1:ipcruisen,:] .= readprop("cruise_TSFC")
     pare[ieTSFC,ipdescent1:ipdescentn,:] .= readprop("descent_TSFC")
 
-else #unrecognized input
-    @warn("The engine type is not recognized")
-end
-
-nac = readprop("Nacelles")
-dnac = dprop["Nacelles"]
-    #- nacelle drag stuff
-    parg[igrSnace] = read_input("nacelle_pylon_wetted_area_ratio", nac, dnac)
-    parg[igrVnace] = read_input("nacelle_local_velocity_ratio", nac, dnac)
-
-    weight = readprop("Weight")
-    dweight = dprop["Weight"]
-        parg[igfeadd] = read_input("engine_access_weight_fraction", weight, dweight)
-        parg[igfpylon] = read_input("pylon_weight_fraction", weight, dweight)
-        
-        #read/check engine weight model options
-       
-        #TODO: reincorporate "pantalone_basic" and "pantalone_adv" for direct-drive turbofans
-        TF_wmodel = read_input("weight_model", weight, dweight)
-        if !(TF_wmodel in ["md", "fitzgerald_basic", "fitzgerald_adv"]) 
-            error("\"$TF_wmodel\" engine weight model was specifed. 
-            Engine weight can only be \"MD\", \"fitzgerald_basic\" or \"fitzgerald_adv\".")
-        end
-    elseif compare_strings(TF_wmodel, "fractional_weight")
-        parg[igfeng] = read_input("engine_weight_fraction", weight, dweight)
-        engineweightname = "fractional_weight"
-        engineweight! = TASOPT.engine.fractional_engine_weight!
-    elseif compare_strings(TF_wmodel, "constant_weight")
-        eng_weight = Force(read_input("engine_weight_total", weight, dweight))
-        engineweightname = "constant_weight"
-        function constant_engine_weight(ac)
-            ac.parg[igWeng] = eng_weight
-        end
-        engineweight! = constant_engine_weight
-    elseif compare_strings(propsys, "te")
-        @warn("Propulsion weight models for turboelectric are currently not available.")
-    end
-
-
[…]
