### 2025-10-02 6baf30a
added calculated CMVf1 functionality with flag (`ac.fuselage.calculates_CMVf1`) and test
