+++ test/unit_test_aero.jl
@@ -270,7 +270,7 @@ end
     Mach = 0.80000000000000004
     fex = 1.0300000000000000
 
-    #Fortran reuslts
+    #Fortran results
     uebl = [0.0000000000000000 0.37728242338296181 0.64847633615516287 0.82695029534888154 0.94403149002962905 1.0255943149422753 1.0865499203351234 1.1365507804257704 1.1206431211650818 1.0440948790037765 1.0240821955777670 1.0158405647282294 1.0117347512309973 1.0095591700001416 1.0085471996187523 1.0083948355769166 1.0090823376403273 1.0108720300588245 1.0145032985652438 1.0221896394842940 1.0402657551827854 1.1059092729914484 1.1091568618457064 1.0543655732241883 0.97490990951276690 0.89368511298240294 0.83118999557894302 0.79419493273373520 0.77809105336189510 0.78142056972012308 0.79242533118539638 0.79772286109406265 0.81053258282277518 0.84123660173182713 0.88716224825623247 0.93244222532354892 0.96275025453959628 0.97871685796354857 0.98673875476417017 0.99107603733351624 0.99366250031367975 0.99532641373569586 0.99645251843379634 0.99724082638968936 0.99780393344085772 0.99820042263719722 0.99838556039722948 0.0000000000000000 0.0000000000000000 0.0000000000000000 0.0000000000000000 0.0000000000000000 0.0000000000000000 0.0000000000000000 0.0000000000000000 0.0000000000000000 0.0000000000000000 0.0000000000000000 0.0000000000000000 0.0000000000000000]
     hkbl = [0.0000000000000000 1.3878006383330910 1.3464796562458770 1.2841329531938646 1.2957711548906594 1.2666086943831041 1.2667873219861920 1.2551499686354026 1.2703003810691702 1.2907324422017787 1.2593018975066568 1.2487806424396475 1.2423297359023246 1.2377736895702194 1.2341982436579075 1.2311759749694335 1.2284089245123413 1.2255808445615450 1.2221794255394689 1.2168451815800081 1.2059592263759851 1.1693954460989937 1.1814318659577450 1.2288851612693361 1.3020077541398134 1.4034862848531129 1.5185526081155398 1.6114897097201484 1.6546827444891374 1.6259707518242759 1.5786779780686202 1.5551481333557295 1.4996868800960224 1.4070761827361435 1.3136726651554176 1.2479933969997437 1.2114726616058633 1.1921668657173372 1.1806347627072433 1.1724280279402921 1.1658003248537754 1.1600708111266333 1.1549504781978848 1.1503025689488813 1.1460563390645413 1.1421769429235715 1.1386910005249200 0.0000000000000000 0.0000000000000000 0.0000000000000000 0.0000000000000000 0.0000000000000000 0.0000000000000000 0.0000000000000000 0.0000000000000000 0.0000000000000000 0.0000000000000000 0.0000000000000000 0.0000000000000000 0.0000000000000000]
[…]
