+++ test/unit_test_io.jl
@@ -87,8 +87,8 @@
     @test size(csv1,1) == 4 #4 rows w default indices
     @test size(csv2,1) == 1 #1 row with addl indices (all)
 
-    @test length(csv1[1]) == 763 # = entries w/ full ac `struct` and in default_output_indices
-    @test length(csv2[1]) == 1133 # = entries w/ full ac `struct` and all output_indices
+    @test length(csv1[1]) == 764 # = entries w/ full ac `struct` and in default_output_indices
+    @test length(csv2[1]) == 1134 # = entries w/ full ac `struct` and all output_indices
 
     #test the nested vectors within par arrays
     #a: row 1 in both csvs matches the design cruise point/mission 
