### 2025-10-02 4f50a35
added calculated CMVf1 functionality with flag (`ac.fuselage.calculates_CMVf1`) and test
