+++ src/data_structs/fuselage.jl
@@ -65,6 +65,8 @@ $TYPEDFIELDS
     weight_frac_skin_addl::Float64 = 0
     """Fuselage Shell Modulus Ratio Ebend/Eskin"""
     ratio_young_mod_fuse_bending::Float64 = 0
+    """Pitching moment estimation method flag"""
+    calculates_CMVf1::Bool = false
 end
 
 function dx_cabin(fuse::Fuselage)
