### 2025-10-02 cbb289a
unrelated tweak: enhanced engine point non-convergence error message/function
