+++ src/engine/turbofan/tfcalc.jl
@@ -677,8 +677,8 @@ end # tfcalc
 
 function check_engine_convergence_failure(pare)
         if sum(pare[ieConvFail, :]) > 0.0 #If any operating point failed to converge
-                return true
-        else
-                return false #All operating points converged
+                warn("Some engine points failed to converge: "*string(pare[ieConvFail,:]))
+        # else
+        #         nothing #All operating points converged
         end
 end
