### 2025-10-14 4d6e42d
added drag breakdown/evolution plotter. edited signature of plot_details()
