### 2025-10-17 1ce8208
added readable names to ip xaxis via `ip_labels` array, generated programmatically in `index.inc`
