### 2025-10-19 fed004f
added deviation points (no deviation implementation
