### 2025-10-19 b883956
initial implementation of deviation
