+++ src/mission/mission_iteration.jl
@@ -435,8 +435,6 @@ function _mission_iteration!(ac, imission, Ldebug; calculate_cruise = false)
             BW = W + para[iaWbuoy, ip]
             F = BW * (DoL + para[iagamV, ip])
             pare[ieFe, ip] = F / parg[igneng] #Store required thrust for engine calcs
-            Wpay = parg[igWpay]
-            
             eng.enginecalc!(ac, "off_design", imission, ip, initializes_engine)
 
       end
@@ -480,103 +478,186 @@ function _mission_iteration!(ac, imission, Ldebug; calculate_cruise = false)
       dRcruise = (alte - altc - gamVdeb * (Rangetot - dRclimb)) / (gamVcr1 - gamVdeb)
 
       altd = altc + gamVcr1 * dRcruise
-      # println("Cruise Fe inside mission", pare[ieFe, ipcruise1])
 
-      # Final cruise point
+      # ---- build cruise/deviation profile between ipcruise1 and ipcruise2
+      Δh_dev = parm[imDeviationDH]
+      R_dev_req = parm[imDeviationDL]
+      S_dev_req = parm[imDeviationStart]
+
+      cruise_indices = (ipcruise1, ipdeviation1, ipdeviation2, ipdeviation3, ipdeviation4, ipcruise2)
+      ncr = length(cruise_indices)
+      # Initializes distance, offsets, climb rate arrays
+      s_values = zeros(Float64, ncr)
+      offsets = zeros(Float64, ncr)
+      segment_gammas = fill(gamVcr1, ncr - 1)
+
+      dR_available = max(dRcruise, 0.0)
+      R_dev = clamp(R_dev_req, 0.0, dR_available)
+
+      # TODO: This is a basic flag for the initial implementation, 
+      # would want it to actually check if start of cruise / length of 
+      #       deviation is shorter than cruise distance etc
+      eps_range = 1.0e-6
+      eps_height = 1.0e-3
+      use_deviation = (R_dev > eps_range) && (abs(Δh_dev) > eps_height)
+
+      if use_deviation
+            γ_climb = para[iagamV, ipclimbn]
+            γ_abs = abs(γ_climb)
+            eps_gamma = 1.0e-6
+            if γ_abs < eps_gamma
+                  use_deviation = false
+            else
+                  sign_h = sign(Δh_dev)
+                  γ_up = sign_h * γ_abs
+                  γ_dn = -sign_h * γ_abs
+                  den_up = γ_up - gamVcr1
+                  den_dn = γ_dn - gamVcr1
+                  eps_den = 1.0e-8
+                  if abs(den_up) < eps_den || abs(den_dn) < eps_den
+                        use_deviation = false
+                  else
+                        coef_total = (1.0 / den_up) - (1.0 / den_dn)
+                        if abs(coef_total) < eps_den
+                              use_deviation = false
+                        else
+                              Δh_limit = R_dev / abs(coef_total)
+                              Δh_eff = clamp(Δh_dev, -Δh_limit, Δh_limit)
+                              if abs(Δh_eff) < eps_height
+                                    use_deviation = false
+                              else
+                                    Δs_up = Δh_eff / den_up
+                                    Δs_dn = -Δh_eff / den_dn
+                                    if (Δs_up < 0.0) || (Δs_dn < 0.0)
[…]
