+++ example/defaults/default_input.toml
@@ -111,8 +111,9 @@
                        # (see `fly_mission!()` and `PayloadRange()` for details, or to recalculate altitude instead).
 
 [Mission.Deviation]
-    deviation_dh = ["0.0", "1000.0 ft"]
-    deviation_dl = ["0.0", "50 nmi"]
+    deviation_dh = ["0.0 ft", "1000.0 ft"]
+    deviation_dl = ["0.0 nmi", "50 nmi"]
+    deviation_start = ["0.0 nmi", "200.0 nmi"]
 
 [Mission.Descent]
     descent_angle_top-of-descent = "-3.0 deg"
