+++ src/mission/mission_iteration.jl
@@ -486,13 +486,15 @@ function _mission_iteration!(ac, imission, Ldebug; calculate_cruise = false)
 
       cruise_indices = (ipcruise1, ipdeviation1, ipdeviation2, ipdeviation3, ipdeviation4, ipcruise2)
       ncr = length(cruise_indices)
-      # Initializes distance, offsets, climb rate arrays
+      # Track cumulative range from cruise start for each waypoint
       s_values = zeros(Float64, ncr)
+      # Altitude offset relative to the nominal cruise-climb line
       offsets = zeros(Float64, ncr)
+      # γ assigned to each cruise segment (piecewise constant between indices)
       segment_gammas = fill(gamVcr1, ncr - 1)
 
       dR_available = max(dRcruise, 0.0)
-      R_dev = clamp(R_dev_req, 0.0, dR_available)
+      R_dev = clamp(R_dev_req, 0.0, dR_available) # cannot exceed remaining cruise
 
       # TODO: This is a basic flag for the initial implementation, 
       # would want it to actually check if start of cruise / length of 
@@ -511,8 +513,8 @@ function _mission_iteration!(ac, imission, Ldebug; calculate_cruise = false)
                   sign_h = sign(Δh_dev)
                   γ_up = sign_h * γ_abs
                   γ_dn = -sign_h * γ_abs
-                  den_up = γ_up - gamVcr1
-                  den_dn = γ_dn - gamVcr1
+                  den_up = γ_up - gamVcr1   # difference between climb γ and baseline γ
+                  den_dn = γ_dn - gamVcr1   # difference between descent γ and baseline γ
                   eps_den = 1.0e-8
                   if abs(den_up) < eps_den || abs(den_dn) < eps_den
                         use_deviation = false
@@ -522,6 +524,7 @@ function _mission_iteration!(ac, imission, Ldebug; calculate_cruise = false)
                               use_deviation = false
                         else
                               Δh_limit = R_dev / abs(coef_total)
+                              # Limit requested altitude change so climb/descent legs stay feasible
                               Δh_eff = clamp(Δh_dev, -Δh_limit, Δh_limit)
                               if abs(Δh_eff) < eps_height
                                     use_deviation = false
@@ -534,8 +537,10 @@ function _mission_iteration!(ac, imission, Ldebug; calculate_cruise = false)
                                           R_hold = max(R_dev - (Δs_up + Δs_dn), 0.0)
                                           R_dev_total = Δs_up + R_hold + Δs_dn
                                           S_dev_max = max(dRcruise - R_dev_total, 0.0)
+                                          # Place deviation entry after requested start distance while staying in cruise span
                                           S_dev = clamp(S_dev_req, 0.0, S_dev_max)
 
+                                          # Assemble cumulative range markers for all cruise nodes
                                           s_values[1] = 0.0
                                           s_values[2] = S_dev
                                           s_values[3] = S_dev + Δs_up
@@ -543,6 +548,7 @@ function _mission_iteration!(ac, imission, Ldebug; calculate_cruise = false)
                                           s_values[5] = s_values[4] + Δs_dn
                                           s_values[6] = dRcruise
 
+                                          # Compose straight–climb–level–descent–straight γ profile
                                           segment_gammas = [gamVcr1, γ_up, gamVcr1, γ_dn, gamVcr1]
                                     end
                               end
@@ -552,6 +558,7 @@ function _mission_iteration!(ac, imission, Ldebug; calculate_cruise = false)
[…]
