### 2025-10-20 b8001f3
Time() -> Times() to deconflict with Dates.Time (#193)
* turn helper fxn Time() into Times() to deconflict with Datetime.jl

* fixed Documenter breaking
