+++ src/IO/read_input.jl
@@ -51,7 +51,7 @@ Density(x)  = convertDensity(parse_unit(x)...)
 Area(x)     = convertArea(parse_unit(x)...)
 Vol(x)      = convertVolume(parse_unit(x)...)
 Angle(x)    = convertAngle(parse_unit(x)...)
-Time(x)     = convertTime(parse_unit(x)...)
+Times(x)     = convertTime(parse_unit(x)...)
 Temp(x)     = convertTemp(parse_unit(x)...)
 
 
@@ -469,8 +469,8 @@ if !(has_wing_fuel) #If fuel is stored in fuselage
 
     fuse_tank.pvent = Pressure(readfuel_storage("pressure_venting"))
     fuse_tank.pinitial = Pressure(readfuel_storage("pressure_initial"))
-    fuse_tank.t_hold_orig = Time(readfuel_storage("hold_departure"))
-    fuse_tank.t_hold_dest = Time(readfuel_storage("hold_arrival"))
+    fuse_tank.t_hold_orig = Times(readfuel_storage("hold_departure"))
+    fuse_tank.t_hold_dest = Times(readfuel_storage("hold_arrival"))
     
     fuse_tank.ftankadd = readfuel_storage("additional_mass_fraction")
     fuse_tank.ew = readfuel_storage("weld_efficiency")
