+++ dev/assets/documenter.js
@@ -33,8 +33,7 @@ requirejs.config({
       "highlight"
     ]
   }
-}
-});
+}});
 ////////////////////////////////////////////////////////////////////////////////
 require(['jquery', 'katex', 'katex-auto-render'], function($, katex, renderMathInElement) {
 $(document).ready(function() {
@@ -65,7 +64,6 @@ $(document).ready(function() {
     "\\Tstd": "T_{\\mathrm{std}}"
   }
 }
-
   );
 })
 
