### 2025-10-22 56073bc
Drag breakdown / evolution plots (#192)
* added drag breakdown/evolution plotter. edited signature of plot_details()

* added readable names to ip xaxis via `ip_labels` array, generated programmatically in `index.inc`

* added tests for `plot_drag_breakdown()`
