+++ .github/workflows/CI.yml
@@ -27,7 +27,7 @@ jobs:
               - x64
     steps:
       - name: Checkout code
-        uses: actions/checkout@v4
+        uses: actions/checkout@v5
 
       - name: Set up Julia
         uses: julia-actions/setup-julia@v2
