### 2025-11-01 459f043
Minor doc string updates as I remind myself what I did here
