### 2025-11-02 f052fc1
added feasibility checks in deviation using available thrust (similar to odperf)
