+++ src/mission/mission_iteration.jl
@@ -481,7 +481,7 @@ function _mission_iteration!(ac, imission, Ldebug; calculate_cruise = false)
 
       # ---- build cruise/deviation profile between ipcruise1 and ipcruise2
       Δh_dev = parm[imDeviationDH]
-      R_dev_req = parm[imDeviationDL]
+      R_dev_req = parm[imDeviationDL] # interpreted as level-leg distance between deviation waypoints 3 and 4
       S_dev_req = parm[imDeviationStart]
 
       cruise_indices = (ipcruise1, ipdeviation1, ipdeviation2, ipdeviation3, ipdeviation4, ipcruise2)
@@ -495,69 +495,118 @@ function _mission_iteration!(ac, imission, Ldebug; calculate_cruise = false)
 
       dR_available = max(dRcruise, 0.0)
       R_dev = clamp(R_dev_req, 0.0, dR_available) # cannot exceed remaining cruise
-
-      # TODO: This is a basic flag for the initial implementation, 
-      # would want it to actually check if start of cruise / length of 
-      #       deviation is shorter than cruise distance etc
       eps_range = 1.0e-6
       eps_height = 1.0e-3
-      use_deviation = (R_dev > eps_range) && (abs(Δh_dev) > eps_height)
+
+      # Evaluate feasibility checks before making deviation profile.
+      use_deviation = false
+      if dR_available > eps_range
+            use_deviation = true
+      elseif R_dev > eps_range
+            use_deviation = true
+      elseif abs(Δh_dev) > eps_height
+            use_deviation = true
+      end
 
       if use_deviation
             γ_climb = para[iagamV, ipclimbn]
             γ_abs = abs(γ_climb)
             eps_gamma = 1.0e-6
             if γ_abs < eps_gamma
+                  if Ldebug
+                        warn("Deviation skipped: reference climb angle at top-of-climb is near zero.")
+                  end
                   use_deviation = false
             else
-                  sign_h = sign(Δh_dev)
-                  γ_up = sign_h * γ_abs
-                  γ_dn = -sign_h * γ_abs
-                  den_up = γ_up - gamVcr1   # difference between climb γ and baseline γ
-                  den_dn = γ_dn - gamVcr1   # difference between descent γ and baseline γ
-                  eps_den = 1.0e-8
-                  if abs(den_up) < eps_den || abs(den_dn) < eps_den
+                  BW_cap = para[iafracW, ipcruise1] * WMTO + para[iaWbuoy, ipcruise1]
+                  # Commented for now but might need since dividing by this later
+                  # BW_cap = max(BW_cap, 1.0e-6)
+                  DoL_cap = para[iaCD, ipcruise1] / para[iaCL, ipcruise1]
+                  # Use the available thrust at top of climb
+                  F_available = max(pare[ieFe, ipclimbn] * parg[igneng], 0.0)
+                  sin_cap = F_available / BW_cap - DoL_cap
+                  if sin_cap <= eps_gamma
+                        if Ldebug
+                              warn("Deviation skipped: available thrust at cruise cannot sustain a climb above the baseline line.")
+                        end
                         use_deviation = false
                   else
-                        coef_total = (1.0 / den_up) - (1.0 / den_dn)
-                        if abs(coef_total) < eps_den
+                        γ_cap = asin(sin_cap)
+                        γ_eff = min(γ_abs, γ_cap)
+                        if γ_eff < eps_gamma
+                              if Ldebug
+                                    warn("Deviation skipped: available thrust leads to deviation climb angle of zero.")
+                              end
                               use_deviation = false
                         else
-                              Δh_limit = R_dev / abs(coef_total)
-                              # Limit requested altitude change so climb/descent legs stay feasible
[…]
