### 2025-11-02 2014210
- renamed `CMVf1` to `pitching_moment_volume` (always within a fuselage reference/object) - edit to slender body theory assumption - tweaked logic, properly assign `fuselage.calculates_pitching_moment_volume`
