+++ example/defaults/default_regional.toml
@@ -174,9 +174,9 @@
     add_payload_weight_fraction = 0.35 # Additional payload proportional Wfrac
 
 [Fuselage.Aero]
-    calculates_CMVf1 = false  # if true, CMVf1 calculated during size_aircraft!() via slender body theory
-                                # if false, value specified by fuse_moment_volume_deriv is used
-    fuse_moment_volume_deriv = "1450.0 ft3" # d(Mfuse/q)/dCL
+    calculates_pitching_moment_volume = false  # if true, CMVf1 calculated during size_aircraft!() via slender body theory
+                                # if false, value specified by pitching_moment_volume is used
+    pitching_moment_volume = "1450.0 ft3" # d(Mfuse/q)/dCL
     CL_zero_fuse_moment      = 0.185         # CL when Mfuse = 0
     #Fuselage velocity overspeed at wing
     wingroot_fuse_overspeed  = 0.0125
