### 2025-11-03 b5b16fb
changed feasibility checks for deviation
