### 2025-11-03 cae3745
added unit test for deviation
