+++ test/runtests.jl
@@ -17,6 +17,7 @@ include(TASOPT.__TASOPTindices__)
     include("unit_test_cryotank.jl")
     include("unit_test_engine.jl")
     include("unit_test_missions.jl")
+    include("unit_test_deviation.jl")
     include("unit_test_electric_machines.jl")
     include("unit_test_outputs.jl")
     include("unit_test_io.jl")
