+++ test/unit_test_deviation.jl
@@ -1,0 +1,277 @@
+@testset "mission deviation" verbose=true begin
+    default_ac = load_default_model()
+    size_aircraft!(default_ac; printiter = false)
+
+    mission_idx = 1
+    cruise_nodes = (
+        TASOPT.ipcruise1,
+        TASOPT.ipdeviation1,
+        TASOPT.ipdeviation2,
+        TASOPT.ipdeviation3,
+        TASOPT.ipdeviation4,
+        TASOPT.ipcruise2,
+    )
+
+    function get_deviation_info(ac, imission)
+        range_start = ac.para[TASOPT.iaRange, TASOPT.ipcruise1, imission]
+        alt_start = ac.para[TASOPT.iaalt, TASOPT.ipcruise1, imission]
+        gam_base = ac.para[TASOPT.iagamV, TASOPT.ipcruise1, imission]
+
+        s_vals = [
+            ac.para[TASOPT.iaRange, ip, imission] - range_start for ip in cruise_nodes
+        ]
+        offsets = [
+            ac.para[TASOPT.iaalt, ip, imission] - (alt_start + gam_base * s)
+            for (ip, s) in zip(cruise_nodes, s_vals)
+        ]
+
+        gam_up = ac.para[TASOPT.iagamV, TASOPT.ipdeviation1, imission]
+        gam_dn = ac.para[TASOPT.iagamV, TASOPT.ipdeviation3, imission]
+
+        base_length = (s_vals[3] - s_vals[2]) + (s_vals[5] - s_vals[4])
+        R_hold = s_vals[4] - s_vals[3]
+
+        return (
+            s_vals = s_vals,
+            offsets = offsets,
+            gam_base = gam_base,
+            gam_up = gam_up,
+            gam_dn = gam_dn,
+            base_length = base_length,
+            R_hold = R_hold,
+            dRcruise = s_vals[end],
+            Δh_eff = offsets[3],
+        )
+    end
+
+    function gamma_cap(ac, imission)
+        WMTO = ac.parg[TASOPT.igWMTO]
+        BW_cap =
+            ac.para[TASOPT.iafracW, TASOPT.ipcruise1, imission] * WMTO +
+            ac.para[TASOPT.iaWbuoy, TASOPT.ipcruise1, imission]
+        DoL_cap =
+            ac.para[TASOPT.iaCD, TASOPT.ipcruise1, imission] /
+            ac.para[TASOPT.iaCL, TASOPT.ipcruise1, imission]
+        F_available =
+            ac.pare[TASOPT.ieFe, TASOPT.ipclimbn, imission] * ac.parg[TASOPT.igneng]
+        sin_cap = F_available / BW_cap - DoL_cap
+        return asin(clamp(sin_cap, -1.0, 1.0))
+    end
+
+    @testset "Sample deviation test" begin
+        ac_default = deepcopy(default_ac)
+
+        ac_default.parm[TASOPT.imDeviationDH, mission_idx] = 304.8 # 1000 ft
+        ac_default.parm[TASOPT.imDeviationDL, mission_idx] = 185200 # 100 nmi
+        ac_default.parm[TASOPT.imDeviationStart, mission_idx] = 185200 # 100 nmi
+
+        TASOPT._mission_iteration!(ac_default, mission_idx, false)
+
+        snap = get_deviation_info(ac_default, mission_idx)
+
+        Δh_req = ac_default.parm[TASOPT.imDeviationDH, mission_idx]
+        R_dev_req = ac_default.parm[TASOPT.imDeviationDL, mission_idx]
[…]
