### 2025-11-03 262a1dd
rewrote mission iteration to use original function for no deviation
