+++ src/mission/mission_iteration.jl
@@ -515,8 +515,6 @@ function _mission_iteration!(ac, imission, Ldebug; calculate_cruise = false)
                   use_deviation = false
             else
                   BW_cap = para[iafracW, ipcruise1] * WMTO + para[iaWbuoy, ipcruise1]
-                  # Commented for now but might need since dividing by this later
-                  # BW_cap = max(BW_cap, 1.0e-6)
                   DoL_cap = para[iaCD, ipcruise1] / para[iaCL, ipcruise1]
                   # Use the available thrust at top of climb
                   F_available = max(pare[ieFe, ipclimbn] * parg[igneng], 0.0)
@@ -602,121 +600,219 @@ function _mission_iteration!(ac, imission, Ldebug; calculate_cruise = false)
                         end
                   end
             end
-      else
-            # No deviation requested, reuse non-deviation linear interpolation between cruise markers
-            for (i, frac) in enumerate(range(0.0, stop = 1.0, length = ncr))
-                  s_values[i] = frac * dRcruise
+      end
+
+      if use_deviation
+            offsets[1] = 0.0
+            for seg = 1:ncr-1
+                  Δs = s_values[seg+1] - s_values[seg]
+                  # Offset integrates local γ minus baseline cruise γ so we return to the base profile
+                  offsets[seg+1] = offsets[seg] + (segment_gammas[seg] - gamVcr1) * Δs
             end
-            segment_gammas .= gamVcr1
-      end
+            offsets[end] = 0.0 # Force return to nominal cruise altitude at ipcruise2
 
-      offsets[1] = 0.0
-      for seg = 1:ncr-1
-            Δs = s_values[seg+1] - s_values[seg]
-            # Offset integrates local γ minus baseline cruise γ so we return to the base profile
-            offsets[seg+1] = offsets[seg] + (segment_gammas[seg] - gamVcr1) * Δs
-      end
-      offsets[end] = 0.0 # Force return to nominal cruise altitude at ipcruise2
+            range_start = para[iaRange, ipcruise1]
+            for (idx, ipx) in enumerate(cruise_indices)
+                  para[iaRange, ipx] = range_start + s_values[idx]
+                  para[iaalt, ipx] = altc + gamVcr1 * s_values[idx] + offsets[idx]
+            end
 
-      range_start = para[iaRange, ipcruise1]
-      for (idx, ipx) in enumerate(cruise_indices)
-            para[iaRange, ipx] = range_start + s_values[idx]
-            para[iaalt, ipx] = altc + gamVcr1 * s_values[idx] + offsets[idx]
-      end
+            for seg = 1:ncr-1
+                  start_idx = cruise_indices[seg]
+                  para[iagamV, start_idx] = segment_gammas[seg]
+            end
+            para[iagamV, cruise_indices[end]] = para[iagamV, cruise_indices[end-1]]
 
-      for seg = 1:ncr-1
-            start_idx = cruise_indices[seg]
-            para[iagamV, start_idx] = segment_gammas[seg]
-      end
-      para[iagamV, cruise_indices[end]] = para[iagamV, cruise_indices[end-1]]
+            Mach_cruise = para[iaMach, ipcruise1]
+            # Recompute ambient conditions, trim, drag, and engine state at every cruise waypoint
[…]
