### 2025-11-04 85ba595
updates to mission iteration - no deviation means same code as old + fix on climb angle update for next point in deviation
