+++ src/mission/mission_iteration.jl
@@ -617,17 +617,11 @@ function _mission_iteration!(ac, imission, Ldebug; calculate_cruise = false)
                   para[iaalt, ipx] = altc + gamVcr1 * s_values[idx] + offsets[idx]
             end
 
-            for seg = 1:ncr-1
-                  start_idx = cruise_indices[seg]
-                  para[iagamV, start_idx] = segment_gammas[seg]
-            end
-            para[iagamV, cruise_indices[end]] = para[iagamV, cruise_indices[end-1]]
-
             Mach_cruise = para[iaMach, ipcruise1]
-            # Recompute ambient conditions, trim, drag, and engine state at every cruise waypoint
-            # so the deviation feeds straight into the existing range/time integration.
-            for k in 1:ncr
-                  ipx = cruise_indices[k]
+
+            function recompute_cruise_point(ipx, γ)
+                  para[iagamV, ipx] = γ
+
                   altkm = para[iaalt, ipx] / 1000.0
                   T0, p0, ρ0, a0, μ0 = atmos(altkm, ΔTatmos)
                   pare[iep0, ipx] = p0
@@ -643,15 +637,13 @@ function _mission_iteration!(ac, imission, Ldebug; calculate_cruise = false)
                   rhocab = max(parg[igpcabin], p0) / (RSL * Tref)
                   para[iaWbuoy, ipx] = (rhocab - ρ0) * gee * parg[igcabVol]
 
-                  γ = para[iagamV, ipx]
                   cosg = cos(γ)
-
                   W = para[iafracW, ipx] * WMTO
                   BW = W + para[iaWbuoy, ipx]
                   Wf = W - Wzero
                   rfuel = Wf / parg[igWfuel]
                   opt_trim_var = "CL_htail"
-                  balance_aircraft!(ac, imission, ipx, rfuel, rpay, ξpay, opt_trim_var; 
+                  balance_aircraft!(ac, imission, ipx, rfuel, rpay, ξpay, opt_trim_var;
                                     Ldebug = Ldebug)
 
                   computes_wing_direct = true
@@ -659,7 +651,7 @@ function _mission_iteration!(ac, imission, Ldebug; calculate_cruise = false)
                   DoL = para[iaCD, ipx] / para[iaCL, ipx]
                   W = para[iafracW, ipx] * WMTO
                   BW = W + para[iaWbuoy, ipx]
-                  Ftotal = BW * (DoL + para[iagamV, ipx])
+                  Ftotal = BW * (DoL + γ)
                   pare[ieFe, ipx] = Ftotal / parg[igneng]
 
                   eng.enginecalc!(ac, "off_design", imission, ipx, initializes_engine)
@@ -667,26 +659,28 @@ function _mission_iteration!(ac, imission, Ldebug; calculate_cruise = false)
                   Ftotal = pare[ieFe, ipx] * parg[igneng]
                   TSFC = pare[ieTSFC, ipx]
                   V = pare[ieu0, ipx]
-                  cosg = cos(para[iagamV, ipx])
                   FoW[ipx] = Ftotal / (BW * cosg) - DoL
                   mfuel = pare[iemfuel, ipx]
                   FFC[ipx] = mfuel * gee / (W * V * cosg)
                   Vgi[ipx] = 1.0 / (V * cosg)
-                  para[iaROC, ipx] = sin(para[iagamV, ipx]) * V * 60.0 / ft_to_m
+                  para[iaROC, ipx] = sin(γ) * V * 60.0 / ft_to_m
+            end
 
-                  if k > 1
-                        # Integrate range, time, and weight fraction over the cruise segment
[…]
