### 2025-11-04 6c9c9be
add easy way to get y and z components of field and control points in WakeSystem
