### 2025-11-06 2cde517
Add trefftz_config to hold discretization settings
