### 2025-11-06 9471006
Add trefftz_config to hold discretization settings
