### 2025-11-06 1a330ed
update and clean up trefftz config.
