+++ src/aero/trefftz_config.jl
@@ -2,7 +2,7 @@
 # related to the Trefftz plane drag estimation method.
 
 """
-Struct to hold the discretization parameters for surface panels 
+Struct to hold the discretization parameters for surface panels
 in the far field drag estimates in the Trefftz Plane.
 """
 struct SurfaceDiscretization
@@ -17,37 +17,92 @@ struct SurfaceDiscretization
     end
 end
 
-function get_trefftz_discretization(s::String)
-    quality = uppercase(s)
-    if quality == "FINE" #328 panels. Reference case. 
-        return (wing_panels = SurfaceDiscretization(160, 48, 24),
-                tail_panels = SurfaceDiscretization(80, 0, 16))
-    elseif quality == "MEDIUM" #gives about 0.60% higher CDi relative to the FINE
-        return (wing_panels = SurfaceDiscretization(40, 12, 6),
-                tail_panels = SurfaceDiscretization(20, 0, 4))
-    elseif quality == "COARSE" # COARSE #gives about 1.8% higher CDi relative to the FINE
-        return (wing_panels = SurfaceDiscretization(20, 6, 3),
-                tail_panels = SurfaceDiscretization(10, 0, 2))
-    else
-        throw(ArgumentError("Invalid discretization resolution: $s. Must be 'COARSE', 'MEDIUM', or 'FINE'."))
-    end
-end
-
-const DEFAULT_TREFFTZ_DISCRETIZATION = get_trefftz_discretization("MEDIUM")
-
 """
-Trefftz plane analysis configuration parameters.
-Controls point clustering and tip correction behavior.
-"""
-struct TrefftzConfig
-    k_tip::Float64          # Tip loading exponent (default: 16.0)
-    bunch::Float64          # Center clustering factor ∈ [0,1] (default: 0.5)
-    root_contraction::Float64  # Root streamline contraction (default: 0.2)
+    TrefftzPlaneConfig
 
-    function TrefftzConfig(; k_tip=16.0, bunch=0.5, root_contraction=0.2)
+Configuration for Trefftz plane induced drag analysis.
+Contains both panel discretization and analysis constants/ parameters.
+
+# Example
+```julia
+config = TrefftzPlaneConfig(
+    wing_panels = SurfaceDiscretization(40, 12, 6),
+    tail_panels = SurfaceDiscretization(20, 0, 4),
+    k_tip = 16.0,
+    bunch = 0.5,
+    root_contraction = 0.2
+)
+```
+"""
+struct TrefftzPlaneConfig
+    wing_panels::SurfaceDiscretization
+    tail_panels::SurfaceDiscretization
+    k_tip::Float64
+    bunch::Float64
+    root_contraction::Float64
+
+    function TrefftzPlaneConfig(wing_panels::SurfaceDiscretization,
+                                tail_panels::SurfaceDiscretization;
+                                k_tip::Float64=16.0,
+                                bunch::Float64=0.5,
+                                root_contraction::Float64=0.2)
         0.0 <= bunch <= 1.0 || throw(ArgumentError("bunch must be in [0,1]"))
         k_tip > 0.0 || throw(ArgumentError("k_tip must be positive"))
         root_contraction > 0.0 || throw(ArgumentError("root_contraction must be positive"))
-        new(k_tip, bunch, root_contraction)
+        
+        new(wing_panels, tail_panels, k_tip, bunch, root_contraction)
     end
 end
+
+"""
+    get_trefftz_config(quality::String; k_tip=16.0, bunch=0.5, root_contraction=0.2)
+
+Create a [`TrefftzPlaneConfig`](@ref).
+
+!!! details "🔃 Inputs and Outputs"
[…]
