### 2025-11-06 63422ab
update default input and read input
