### 2025-11-06 499fea9
update default input and read input
