+++ src/IO/read_input.jl
@@ -191,6 +191,16 @@ if !(typeof(engloc) <: AbstractString && engloc in ["wing", "fuselage"])
         > \"fuselage\" - engines on aft fuselage")
 end
 
+trefftz_resolution_str = read_input("trefftz_resolution", options, doptions)
+trefftz_k_tip = read_input("trefftz_k_tip", options, doptions)
+trefftz_bunch = read_input("trefftz_bunch", options, doptions)
+trefftz_root_contraction = read_input("trefftz_root_contraction", options, doptions)
+TREFFTZ_CONFIG = aerodynamics.get_trefftz_config(trefftz_resolution_str;
+                                                  k_tip=trefftz_k_tip,
+                                                  bunch=trefftz_bunch,
+                                                  root_contraction=trefftz_root_contraction)
+
+
 # Fuel related options
 fuel = read_input("Fuel", data, default)
 dfuel = default["Fuel"]
@@ -1205,7 +1215,9 @@ ac_options = TASOPT.Options(
     
     is_doubledecker = is_doubledecker,
 
-    opt_move_wing = opt_move_wing
+    opt_move_wing = opt_move_wing,
+
+    trefftz_config = TREFFTZ_CONFIG
 )
     
 #Create aircraft object
