### 2025-11-06 f299935
include trefftz_config in aero.
