### 2025-11-06 c90a3b5
Update trefftz calcs to take in config struct and update call sites.
