+++ src/aero/induced_drag.jl
@@ -1,5 +1,5 @@
 """
-      induced_drag!(para, wing, htail)
+      induced_drag!(para, wing, htail, trefftz_config)
 
 Computes the induced drag via the Trefftz plane. Calls [`_trefftz_analysis`](@ref). Formerly, `cditrp!()`.
 
@@ -8,6 +8,7 @@ Computes the induced drag via the Trefftz plane. Calls [`_trefftz_analysis`](@re
       - `para::AbstractArray{Float64}`: Array of `aircraft` model aerodynamic parameters.
       - `wing::TASOPT.Wing`: Wing Structure.
       - `htail::TASOPT.Tail`: Htail Structure.
+      - `trefftz_config::TrefftzPlaneConfig`: Trefftz plane analysis configuration (panel discretization and analysis parameters).
 
       **Outputs:**
       - No explicit outputs. Computed induced drag value and span efficiency are saved to `para` of `aircraft` model.
@@ -16,7 +17,7 @@ Computes the induced drag via the Trefftz plane. Calls [`_trefftz_analysis`](@re
       In an upcoming revision, an `aircraft` `struct` and auxiliary indices will be passed in lieu of pre-sliced `par` arrays.
 
 """
-function induced_drag!(para, wing, htail)
+function induced_drag!(para, wing, htail, trefftz_config::TrefftzPlaneConfig)
 
       CL = para[iaCL]
 
@@ -45,10 +46,6 @@ function induced_drag!(para, wing, htail)
       po       = zeros(Float64, 2)
       CLsurfsp = zeros(Float64, 2)
 
-      npout = zeros(Float64, 2) # outer panel
-      npinn = zeros(Float64, 2) # inner panel
-      npimg = zeros(Float64, 2) # image inside fuselage
-
       #Alternatively can define as b  = [parg[igb], parg[igbh]] for both wing and tail simultaneously 
 #---- wing wake parameters
       fLo =  wing.fuse_lift_carryover
@@ -60,7 +57,7 @@ function induced_drag!(para, wing, htail)
       bo[1] = wing.layout.root_span
 
 #---- span of wing-root streamline in Trefftz Plane
-      bop[1] = wing.layout.root_span * 0.2
+      bop[1] = wing.layout.root_span * trefftz_config.root_contraction
 
       zcent[1]  = wing.layout.z
       gammas[1] = wing.inboard.λ*para[iarcls]
@@ -89,54 +86,10 @@ function induced_drag!(para, wing, htail)
 #---- number of surfaces  (wing, horizontal tail)
       nsurf = 2
 
-#---- number of spanwise intervals
-# -----------------------------------------------------------------------
-# The below # of panels (43 total) gives about 1.8% higher CDi relative to the 
-# finest one here with ~360 panels
-      npout[1] = 20  # outer panel
-      npinn[1] = 6   # inner panel
-      npimg[1] = 3   # image inside fuselage
-  
-      npout[2] = 10  # outer panel
-      npinn[2] = 0   # inner panel
-      if (bo[2] == 0.0) 
-       npimg[2] = 0
-      else
-       npimg[2] = 2   # image inside fuselage  (or inner panel if T-tail)
-      end
-# -----------------------------------------------------------------------
-# The below # of panels (84 total) gives about 0.60% higher CDi relative to the 
-# finest one here with ~360 panels
-#     npout[1] = 40  # outer panel
-#     npinn[1] = 12  # inner panel
[…]
