### 2025-11-06 714727b
Update trefftz calcs to take in config struct and update call sites.
