### 2025-11-06 4ef179b
update options to hold trefftz config
