### 2025-11-06 46c7a92
update aero test call site
