### 2025-11-06 348a07d
Set resolution to COARSE to match prior tests.
