### 2025-11-06 8c91927
Set resolution to COARSE to match prior tests.
