+++ example/defaults/default_input.toml
@@ -46,7 +46,7 @@
 
     prop_sys_arch = "TF" #Options are TF: Turbofan
                          #            constant_TSFC: simple engine with constant TSFC
-    trefftz_resolution = "MEDIUM"
+    trefftz_resolution = "COARSE"
     # Trefftz plane discretization quality
     # "COARSE":  ~1.8% error rel. reference,  43 panels (fast, mainly for quick studies)
     # "MEDIUM":  ~0.6% error rel. reference,  84 panels (balanced, recommended default)
