### 2025-11-07 c7cee4b
remove stale code and add bunch transofrm functions
