+++ src/aero/induced_drag.jl
@@ -101,90 +101,71 @@ function induced_drag!(para, wing, htail, trefftz_config::TrefftzPlaneConfig)
 end # induced_drag!
 
 
-# Trefftz analysis functions -------------------
-# Investigating potential to have wrappers to make code cleaner
-struct tfp_workarrays
-  t::Vector{Float64}
-  y::Vector{Float64}
-  yp::Vector{Float64}
-  z::Vector{Float64}
-  zp::Vector{Float64}
-  gw::Vector{Float64}
-  yc::Vector{Float64}
-  ycp::Vector{Float64}
-  zc::Vector{Float64}
-  zcp::Vector{Float64}
-  gc::Vector{Float64}
-  vc::Vector{Float64}
-  wc::Vector{Float64}
-  vnc::Vector{Float64}
-end
+"""
+    bunch_transform(t, bunch)
 
-struct tfp
-  nsurf::Int
-  npout::Vector{Int}
-  npinn::Vector{Int}
-  npimg::Vector{Int}
-  Sref::Float64
-  bref::Float64
-  b::Vector{Float64}
-  bs::Vector{Float64}
-  bo::Vector{Float64}
-  bop::Vector{Float64}
-  zcent::Vector{Float64}
-  po::Vector{Float64}
-  γt::Vector{Float64}
-  γs::Vector{Float64}
-  CLsurfsp::Vector{Float64}
-  ktip::Float64
-  workarrays::tfp_workarrays
-end
+Apply bunching transformation to parameter `t ∈ [0,1]`.
 
-function trefftz(tf::tfp)
-  
-  _trefftz_analysis(tf.nsurf, tf.npout, tf.npinn, tf.npimg,
-	tf.Sref, tf.bref,
-	tf.b, tf.bs, tf.bo, tf.bop, tf.zcent,
-	tf.po,tf.γt,tf.γs, 1, tf.ktip,
-	true,tf.CLsurfsp,
-  tf.workarrays.t, tf.workarrays.y, tf.workarrays.yp, tf.workarrays.z, tf.workarrays.zp, 
-  tf.workarrays.gw, tf.workarrays.yc, tf.workarrays.ycp, tf.workarrays.zc, tf.workarrays.zcp,
-  tf.workarrays.gc, tf.workarrays.vc, tf.workarrays.wc, tf.workarrays.vnc)
-  ## Note: do not do the following. This will slow the code down. Interestingly seems to have more allocations than the above.
-  # [getproperty(tf.workarrays,f) for f in fieldnames(typeof(tf.workarrays))]...)
+Transforms `t` to cluster points near the center (t=0.5) when `bunch > 0`.
+The transformation is: `t_bunched = t + bunch * t * (1 - t)`
 
+!!! details "🔃 Inputs and Outputs"
+    **Inputs:**
+    - `t::Float64`: Original parameter value ∈ [0,1].
+    - `bunch::Float64`: Clustering factor ∈ [0,1]. Higher values cluster more points near center.
+
+    **Outputs:**
+    - `t_bunched::Float64`: Transformed parameter value.
+
+See also [`inv_bunch_transform`](@ref).
+"""
+@inline function bunch_transform(t, bunch)
+    return t + bunch * t * (1.0 - t)
 end
 
 """
-    _trefftz_analysis(nsurf, npout, npinn, npimg, 
-          Sref, bref, b, bs, bo, bop, 
-          zcent, po, gammat, gammas, 
-          fLo,ktip, specifies_CL, CLsurfsp)
+    inv_bunch_transform(t_bunched, bunch)
+
+Inverse of the bunching transformation. Recovers original `t` from bunched value.
+
+Solves `t_bunched = t + bunch * t * (1 - t)` for `t` using the quadratic formula.
+See also [`bunch_transform`](@ref).
+"""
[…]
