### 2025-11-07 4232964
remove duplicate ref. in docs
try to fix docs

escape [] in md

fix [] again

to squash

update bunch transform docstring
