### 2025-11-07 48207ad
remove duplicate ref. in docs
try to fix docs

escape [] in md

fix [] again

to squash

update bunch transform docstring
