### 2025-11-07 5a3c806
First cut at refactroring Trefftz Plane calcs. NO major speed up yet.
