+++ src/aero/induced_drag.jl
@@ -36,30 +36,14 @@ function induced_drag!(para, wing, htail, trefftz_config::TrefftzPlaneConfig)
 
       specifies_CL = true
 
-      b        = zeros(Float64, 2)
-      bs       = zeros(Float64, 2)
-      bo       = zeros(Float64, 2)
-      bop      = zeros(Float64, 2)
-      zcent    = zeros(Float64, 2)
       gammas   = zeros(Float64, 2)
       gammat   = zeros(Float64, 2)
       po       = zeros(Float64, 2)
       CLsurfsp = zeros(Float64, 2)
 
-      #Alternatively can define as b  = [parg[igb], parg[igbh]] for both wing and tail simultaneously 
 #---- wing wake parameters
       fLo =  wing.fuse_lift_carryover
-#      fLo = 0.0
 
-#---- span, wing-break span, wing-root span
-      b[1]  = wing.layout.span
-      bs[1] = wing.layout.break_span
-      bo[1] = wing.layout.root_span
-
-#---- span of wing-root streamline in Trefftz Plane
-      bop[1] = wing.layout.root_span * trefftz_config.root_contraction
-
-      zcent[1]  = wing.layout.z
       gammas[1] = wing.inboard.λ*para[iarcls]
       gammat[1] = wing.outboard.λ*para[iarclt]
       po[1]     = 1.0
@@ -71,12 +55,6 @@ function induced_drag!(para, wing, htail, trefftz_config::TrefftzPlaneConfig)
       fdut = para[iafdut]
 
 #---- horizontal tail wake parameters
-      b[2]   = htail.layout.span
-      bs[2]  = htail.layout.root_span
-      bo[2]  = htail.layout.root_span
-      bop[2] = htail.layout.root_span
-
-      zcent[2] = htail.layout.z
       gammas[2] = 1.0
       gammat[2] = htail.outboard.λ
       po[2]     = 1.0
@@ -87,8 +65,8 @@ function induced_drag!(para, wing, htail, trefftz_config::TrefftzPlaneConfig)
       nsurf = 2
 
       CLsurf, CLtp, CDtp, sefftp = _trefftz_analysis(nsurf, trefftz_config,
+	wing, htail,
 	Sref, bref,
-	b,bs,bo,bop, zcent,
 	po,gammat,gammas, fLo,
       specifies_CL,CLsurfsp, t, y, yp, z, zp, gw, yc, ycp, zc, zcp, gc, vc, wc, vnc)
       
@@ -135,9 +113,229 @@ See also [`bunch_transform`](@ref).
     return (1.0 + bunch - sqrt((1.0 + bunch)^2 - 4.0 * bunch * t_bunched)) * 0.5 / bunch
 end
 
+@enum WAKE_CONTRACTION_TYPE begin
+    FUSEWAKE
+    WINGWAKE
+end
+
+function get_wake_contraction(wake_type::WAKE_CONTRACTION_TYPE, y, yo, yop)
+      if wake_type === FUSEWAKE
+            # Power law to contract the stream tubes in the fuselage region.
+            @assert y ≤ yo "y must be within fuselage region (y ≤ yo)"
+            yexp = (yo / yop)^2
+            return yop * (y / yo)^yexp
+      else
+            # Mass conservation based contraction outside fuselage region.
+            @assert y ≥ yo "y must be outside fuselage region (y ≥ yo)"
+            return sqrt(y^2 - yo^2 + yop^2)
+      end
+end
+
 """
-    _trefftz_analysis(nsurf, trefftz_config, Sref, bref, b, bs, bo, bop,
-          zcent, po, gammat, gammas, fLo, specifies_CL, CLsurfsp,
+    generate_panel_points!(
[…]
