### 2025-11-07 c4abf00
Factor out all the work arrays into a neat struct.
fix function call signs and update tests.
