+++ src/aero/induced_drag.jl
@@ -68,8 +68,8 @@ function induced_drag!(para, wing, htail, trefftz_config::TrefftzPlaneConfig)
 	wing, htail,
 	Sref, bref,
 	po,gammat,gammas, fLo,
-      specifies_CL,CLsurfsp, t, y, yp, z, zp, gw, yc, ycp, zc, zcp, gc, vc, wc, vnc)
-      
+      specifies_CL, CLsurfsp, TREFFTZ_GEOM, gw, vc, wc, vnc)
+
       # println("$CLsurf, $CLtp, $CDtp, $sefftp")
 
       para[iaCDi] = CDtp
@@ -121,19 +121,25 @@ end
 function get_wake_contraction(wake_type::WAKE_CONTRACTION_TYPE, y, yo, yop)
       if wake_type === FUSEWAKE
             # Power law to contract the stream tubes in the fuselage region.
-            @assert y ≤ yo "y must be within fuselage region (y ≤ yo)"
+            # Use tolerance for boundary case (y ≈ yo)
+            if y > yo * (1.0 + 1e-10)
+                  error("y=$y must be within fuselage region (y ≤ yo=$yo)")
+            end
             yexp = (yo / yop)^2
             return yop * (y / yo)^yexp
       else
             # Mass conservation based contraction outside fuselage region.
-            @assert y ≥ yo "y must be outside fuselage region (y ≥ yo)"
+            # Use tolerance for boundary case (y ≈ yo)
+            if y < yo * (1.0 - 1e-10)
+                  error("y=$y must be outside fuselage region (y ≥ yo=$yo)")
+            end
             return sqrt(y^2 - yo^2 + yop^2)
       end
 end
 
 """
     generate_panel_points!(
-        i, t, y, yp, z, zp, yc, ycp, zc, zcp, gc,
+        i, geom,
         k_start, k_end,
         t_start, t_end,
         e_start, e_end,
@@ -149,7 +155,7 @@ Type-stable and efficient for repeated use in different panel sections.
 """
 @inline function generate_panel_points!(
     i::Int,
-    t, y, yp, z, zp, yc, ycp, zc, zcp, gc,
+    geom::TrefftzGeometry,
     k_start::Int, k_end::Int,
     t_start::Float64, t_end::Float64,
     e_start::Float64, e_end::Float64,
@@ -157,7 +163,7 @@ Type-stable and efficient for repeated use in different panel sections.
     z_start::Float64, z_end::Float64,
     g_start::Float64, g_end::Float64,
     yo::Float64, yop::Float64,
-    bunch::Float64, ktip::Float64, 
+    bunch::Float64, ktip::Float64,
     wake_contract::WAKE_CONTRACTION_TYPE)
 
     @inbounds for k = k_start+1:k_end
@@ -167,11 +173,11 @@ Type-stable and efficient for repeated use in different panel sections.
         fk = (k - k_start) / (k_end - k_start)
 
         # Theta values for field point and control point
-        t[i] = t_start * (1.0 - fk) + t_end * fk
-        tc = 0.5 * (t[i-1] + t[i])
+        geom.t[i] = t_start * (1.0 - fk) + t_end * fk
+        tc = 0.5 * (geom.t[i-1] + geom.t[i])
 
         # Eta values (spanwise position)
-        e  = cos(0.5π * bunch_transform(t[i], bunch))
+        e  = cos(0.5π * bunch_transform(geom.t[i], bunch))
         ec = cos(0.5π * bunch_transform(tc, bunch))
 
         # Interpolation fractions in eta space
[…]
