### 2025-11-19 70956e6
Further refactoring of trefftz calcs
