+++ src/aero/induced_drag.jl
@@ -1,7 +1,8 @@
 """
       induced_drag!(para, wing, htail, trefftz_config)
 
-Computes the induced drag via the Trefftz plane. Calls [`_trefftz_analysis`](@ref). Formerly, `cditrp!()`.
+Computes the induced drag via the Trefftz plane. Calls [`_trefftz_analysis`](@ref). 
+Formerly, `cditrp!()` but now broken up into smaller functions.
 
 !!! details "🔃 Inputs and Outputs"
       **Inputs:**
@@ -27,7 +28,7 @@ function induced_drag!(para, wing, htail, trefftz_config::TrefftzPlaneConfig)
        return
       end
 
-      CLhtail = para[iaCLh]*htail.layout.S/wing.layout.S
+      CLhtail = para[iaCLh]*htail.layout.S/wing.layout.S #normalize by wing area
       # println("CLhtail: $(para[iaCLh]) $(htail.layout.S) $(parg[igS])")
       bref = wing.layout.span
       Sref = wing.layout.S
@@ -44,8 +45,8 @@ function induced_drag!(para, wing, htail, trefftz_config::TrefftzPlaneConfig)
 #---- wing wake parameters
       fLo =  wing.fuse_lift_carryover
 
-      gammas[1] = wing.inboard.λ*para[iarcls]
-      gammat[1] = wing.outboard.λ*para[iarclt]
+      gammas[1] = wing.inboard.λ * para[iarcls]
+      gammat[1] = wing.outboard.λ * para[iarclt]
       po[1]     = 1.0
       CLsurfsp[1] = CL - CLhtail
 
@@ -64,11 +65,11 @@ function induced_drag!(para, wing, htail, trefftz_config::TrefftzPlaneConfig)
 #---- number of surfaces  (wing, horizontal tail)
       nsurf = 2
 
-      CLsurf, CLtp, CDtp, sefftp = _trefftz_analysis(nsurf, trefftz_config,
-	wing, htail,
-	Sref, bref,
-	po,gammat,gammas, fLo,
-      specifies_CL, CLsurfsp, TREFFTZ_GEOM, gw, vc, wc, vnc)
+    CLsurf, CLtp, CDtp, sefftp = _trefftz_analysis(nsurf, trefftz_config,
+        wing, htail,
+        Sref, bref,
+        po, gammat, gammas, fLo,
+        specifies_CL, CLsurfsp, TREFFTZ_GEOM, gw, vc, wc, vnc)
 
       # println("$CLsurf, $CLtp, $CDtp, $sefftp")
 
@@ -111,6 +112,78 @@ See also [`bunch_transform`](@ref).
 """
 @inline function inv_bunch_transform(t_bunched, bunch)
     return (1.0 + bunch - sqrt((1.0 + bunch)^2 - 4.0 * bunch * t_bunched)) * 0.5 / bunch
+end
+
+"""
+    calculate_wake_circulation!(gw, gc, ifrst, ilast, nsurf)
+
+Compute wake vortex circulation from bound circulation `γ_wake = -dΓ/dy`.
+
+At each panel edge, the change in bound circulation is shed into the wake:
+- Root: `gw = 0` (no upstream circulation at the center of the fuselage)
+- Interior: `gw[i] = gc[i-1] - gc[i]` 
+- Tip: `gw[i] = gc[i-1]` 
+"""
+@inline function calculate_wake_circulation!(
+    gw::AbstractVector{Float64},
+    gc::AbstractVector{Float64},
+    i_first::AbstractVector{Int},
+    i_last::AbstractVector{Int},
+    nsurf::Int
+)
+    @inbounds for isurf = 1:nsurf
+        i = i_first[isurf]
[…]
