### 2025-11-19 7400d8b
Further refactoring of trefftz calcs
