### 2025-11-19 089f3df
CLsurfsp should be a vector (was being treated as a matrix)
