+++ test/unit_test_aero.jl
@@ -322,7 +322,7 @@ end
     gammas = [0.77000000000000002,  1.0000000000000000]
     fLo = -0.29999999999999999
     specifies_CL = true
-    CLsurfsp = [1.2502595055642693 1.1976022033901442E-002]
+    CLsurfsp = [1.2502595055642693, 1.1976022033901442E-002]
 
     fort_CLsurf = [1.2502595055642693, 1.1976022033901442E-002]
     fort_CLtp = 1.2622355275981709
