### 2025-11-19 f268d5a
update defaults for tail contraction
