### 2025-11-19 b1c57c3
update defaults for tail contraction
