### 2025-11-26 ee1b5bd
fix the weird exits and replace with errors
