### 2025-11-27 1fe3e41
Helper function in trefftz
