### 2025-11-27 297213e
helper functions for getting panel indices
