### 2025-11-27 d4d5c93
helper functions for getting panel indices
