### 2025-11-27 a4455c7
move geom check to config constructor
