### 2025-11-27 dccf36e
move geom check to config constructor
