### 2025-11-27 52aa5ff
refactor geometry construction into separate function
