### 2025-11-27 de6a619
initial hash logic for Trefftz plane geometry
formatting and small changes

udpate test signature
