+++ src/aero/induced_drag.jl
@@ -20,65 +20,76 @@ Formerly, `cditrp!()` but now broken up into smaller functions.
 """
 function induced_drag!(para, wing, htail, trefftz_config::TrefftzPlaneConfig)
 
-      CL = para[iaCL]
+    CL = para[iaCL]
 
-      if (CL == 0.0) 
-	    para[iaCDi]  = 0.
-	    para[iaspaneff] = 1.0
-       return
-      end
+    if (CL == 0.0)
+        para[iaCDi] = 0.
+        para[iaspaneff] = 1.0
+        return
+    end
 
-      CLhtail = para[iaCLh]*htail.layout.S/wing.layout.S #normalize by wing area
-      # println("CLhtail: $(para[iaCLh]) $(htail.layout.S) $(parg[igS])")
-      bref = wing.layout.span
-      Sref = wing.layout.S
+    CLhtail = para[iaCLh] * htail.layout.S / wing.layout.S #normalize by wing area
+    # println("CLhtail: $(para[iaCLh]) $(htail.layout.S) $(parg[igS])")
+    bref = wing.layout.span
+    Sref = wing.layout.S
 
-      Mach = para[iaMach]
+    Mach = para[iaMach]
 
-      specifies_CL = true
+    specifies_CL = true
 
-      gammas   = zeros(Float64, 2)
-      gammat   = zeros(Float64, 2)
-      po       = zeros(Float64, 2)
-      CLsurfsp = zeros(Float64, 2)
+    gammas = zeros(Float64, 2)
+    gammat = zeros(Float64, 2)
+    po = zeros(Float64, 2)
+    CLsurfsp = zeros(Float64, 2)
 
-#---- wing wake parameters
-      fLo =  wing.fuse_lift_carryover
+    #---- wing wake parameters
+    fLo = wing.fuse_lift_carryover
 
-      gammas[1] = wing.inboard.λ * para[iarcls]
-      gammat[1] = wing.outboard.λ * para[iarclt]
-      po[1]     = 1.0
-      CLsurfsp[1] = CL - CLhtail
+    gammas[1] = wing.inboard.λ * para[iarcls]
+    gammat[1] = wing.outboard.λ * para[iarclt]
+    po[1] = 1.0
+    CLsurfsp[1] = CL - CLhtail
 
-#---- velocity-change fractions at wing spanwise locations due to fuselage flow
-      fduo = para[iafduo]
-      fdus = para[iafdus]
-      fdut = para[iafdut]
+    #---- velocity-change fractions at wing spanwise locations due to fuselage flow
+    fduo = para[iafduo]
+    fdus = para[iafdus]
+    fdut = para[iafdut]
 
-#---- horizontal tail wake parameters
-      gammas[2] = 1.0
-      gammat[2] = htail.outboard.λ
-      po[2]     = 1.0
-      CLsurfsp[2] = CLhtail
+    #---- horizontal tail wake parameters
+    gammas[2] = 1.0
+    gammat[2] = htail.outboard.λ
+    po[2] = 1.0
+    CLsurfsp[2] = CLhtail
 
 
-#---- number of surfaces  (wing, horizontal tail)
-      nsurf = 2
-      
+    #---- number of surfaces  (wing, horizontal tail)
+    nsurf = 2
+
     CLsurf, CLtp, CDtp, sefftp = _trefftz_analysis(nsurf, trefftz_config,
         wing, htail,
         Sref, bref,
         po, gammat, gammas, fLo,
-        specifies_CL, CLsurfsp, TREFFTZ_GEOM, gw, vc, wc, vnc)
+        specifies_CL, CLsurfsp, TREFFTZ_GEOM)
 
-      # println("$CLsurf, $CLtp, $CDtp, $sefftp")
[…]
