### 2025-11-27 b637cec
simplify function signature and initial draft of checking trefftz geometry validity
