+++ src/aero/induced_drag.jl
@@ -475,6 +475,13 @@ function _build_trefftz_geometry!(wing, htail, po, gammat, gammas,
         TREFFTZ_GEOM.zc[i] = 0.0
         TREFFTZ_GEOM.gc[i] = 0.0
     end # nsurf loop
+    # ilast = i_last_tail(trefftz_config) # could use this instead of accumulating i...
+    WS = generate_wake_system(
+        TREFFTZ_GEOM.yp[1:i],
+        TREFFTZ_GEOM.zp[1:i],
+        TREFFTZ_GEOM.ycp[1:i-1],
+        TREFFTZ_GEOM.zcp[1:i-1])
+    return WS
 end
 
 
@@ -515,7 +522,7 @@ function _trefftz_analysis(nsurf, trefftz_config::TrefftzPlaneConfig,
 	wing, htail,
 	Sref, bref,
 	po, gammat, gammas, fLo,
-	specifies_CL, CLsurfsp, geom::TrefftzGeometry, gw, vc, wc, vnc)
+	specifies_CL, CLsurfsp, geom::TrefftzGeometry)
 
       ifrst = zeros(Int, nsurf)
       ilast = zeros(Int, nsurf)
