### 2025-12-03 b6b0588
Make aircraft parametric and hold wake system. This lets us use ensure_trefftz_current! logic
fix the missing export in aero.jl and update the test for # of entries in output

update cache logic
