+++ src/aero/aero.jl
@@ -11,7 +11,7 @@ using ..atmosphere
 using ..TASOPT: __TASOPTindices__, __TASOPTroot__, compare_strings, balance_aircraft!
 
 export airfoil, aircraft_drag!, wing_CM, set_wing_geometry!, wing_loading, tail_loading!, wing_section_cls, 
-    fuselage_drag!, broadcast_fuselage_drag!, plot_airf, aeroperf_sweep
+    fuselage_drag!, broadcast_fuselage_drag!, plot_airf, aeroperf_sweep, WakeSystem
 
 # Define the __init__ function
 #This function gets executed automatically when the module is loaded
