### 2025-12-03 4728a89
Make aircraft parametric and hold wake system. This lets us use ensure_trefftz_current! logic
fix the missing export in aero.jl and update the test for # of entries in output

update cache logic
