### 2025-12-08 8a1208c
read input update for changed deviation var names
