### 2025-12-08 a50fd82
deviation test update for changed var names
