+++ test/unit_test_deviation.jl
@@ -61,17 +61,17 @@
     @testset "Sample deviation test" begin
         ac_default = deepcopy(default_ac)
 
-        ac_default.parm[TASOPT.imDeviationDH, mission_idx] = 304.8 # 1000 ft
-        ac_default.parm[TASOPT.imDeviationDL, mission_idx] = 185200 # 100 nmi
-        ac_default.parm[TASOPT.imDeviationStart, mission_idx] = 185200 # 100 nmi
+        ac_default.parm[TASOPT.imDeviationHeight, mission_idx] = 304.8 # 1000 ft
+        ac_default.parm[TASOPT.imDeviationLength, mission_idx] = 185200 # 100 nmi
+        ac_default.parm[TASOPT.imDeviationStartFromTOC, mission_idx] = 185200 # 100 nmi
 
         TASOPT._mission_iteration!(ac_default, mission_idx, false)
 
         snap = get_deviation_info(ac_default, mission_idx)
 
-        Δh_req = ac_default.parm[TASOPT.imDeviationDH, mission_idx]
-        R_dev_req = ac_default.parm[TASOPT.imDeviationDL, mission_idx]
-        S_dev_req = ac_default.parm[TASOPT.imDeviationStart, mission_idx]
+        Δh_req = ac_default.parm[TASOPT.imDeviationHeight, mission_idx]
+        R_dev_req = ac_default.parm[TASOPT.imDeviationLength, mission_idx]
+        S_dev_req = ac_default.parm[TASOPT.imDeviationStartFromTOC, mission_idx]
 
         gam_base = snap.gam_base
         gam_up = snap.gam_up
@@ -123,9 +123,9 @@
 
         snap = get_deviation_info(ac_default, mission_idx)
 
-        Δh_req = ac_default.parm[TASOPT.imDeviationDH, mission_idx]
-        R_dev_req = ac_default.parm[TASOPT.imDeviationDL, mission_idx]
-        S_dev_req = ac_default.parm[TASOPT.imDeviationStart, mission_idx]
+        Δh_req = ac_default.parm[TASOPT.imDeviationHeight, mission_idx]
+        R_dev_req = ac_default.parm[TASOPT.imDeviationLength, mission_idx]
+        S_dev_req = ac_default.parm[TASOPT.imDeviationStartFromTOC, mission_idx]
 
         gam_base = snap.gam_base
         gam_up = snap.gam_up
@@ -175,16 +175,16 @@
 
     @testset "Deviation start clamps to cruise span" begin
         ac_start = deepcopy(default_ac)
-        ac_start.parm[TASOPT.imDeviationDH, mission_idx] = 304.8 # ensure deviation active
-        ac_start.parm[TASOPT.imDeviationDL, mission_idx] = 185200.0
-        ac_start.parm[TASOPT.imDeviationStart, mission_idx] = 1.0e7
+        ac_start.parm[TASOPT.imDeviationHeight, mission_idx] = 304.8 # ensure deviation active
+        ac_start.parm[TASOPT.imDeviationLength, mission_idx] = 185200.0
+        ac_start.parm[TASOPT.imDeviationStartFromTOC, mission_idx] = 1.0e7
 
         TASOPT._mission_iteration!(ac_start, mission_idx, false)
         snap = get_deviation_info(ac_start, mission_idx)
 
-        S_dev_req = ac_start.parm[TASOPT.imDeviationStart, mission_idx]
-        R_dev_req = ac_start.parm[TASOPT.imDeviationDL, mission_idx]
-        Δh_req = ac_start.parm[TASOPT.imDeviationDH, mission_idx]
[…]
