### 2025-12-08 5c57766
updated deviation logic after prashanth review, and other changes after personal review
