+++ src/mission/mission_iteration.jl
@@ -480,207 +480,292 @@ function _mission_iteration!(ac, imission, Ldebug; calculate_cruise = false)
       altd = altc + gamVcr1 * dRcruise
 
       # ---- build cruise/deviation profile between ipcruise1 and ipcruise2
-      Δh_dev = parm[imDeviationDH]
-      R_dev_req = parm[imDeviationDL] # interpreted as level-leg distance between deviation waypoints 3 and 4
-      S_dev_req = parm[imDeviationStart]
+      # Inputs describing desired deviation geometry
+      Δh_dev = parm[imDeviationHeight]
+      R_dev_req = parm[imDeviationLength] # interpreted as level-leg distance between deviation waypoints 3 and 4
+      S_dev_req = parm[imDeviationStartFromTOC]
 
       cruise_indices = (ipcruise1, ipdeviation1, ipdeviation2, ipdeviation3, ipdeviation4, ipcruise2)
       ncr = length(cruise_indices)
-      # Track cumulative range from cruise start for each waypoint
-      s_values = zeros(Float64, ncr)
-      # Altitude offset relative to the nominal cruise-climb line
-      offsets = zeros(Float64, ncr)
-      # γ assigned to each cruise segment (piecewise constant between indices)
-      segment_gammas = fill(gamVcr1, ncr - 1)
 
+      # Available cruise distance and numerical cutoffs for range/height
       dR_available = max(dRcruise, 0.0)
-      R_dev = clamp(R_dev_req, 0.0, dR_available) # cannot exceed remaining cruise
       eps_range = 1.0e-6
       eps_height = 1.0e-3
 
-      # Evaluate feasibility checks before making deviation profile.
-      use_deviation = false
-      if abs(Δh_dev) > eps_height && dR_available > eps_range && R_dev > eps_range
-            use_deviation = true
+      # TODO: Right now parm stores Float64, thus this use_deviation is 1.0 or 0.0, we need to just make a Mission object instead
+      use_deviation = parm[imUseDeviation] != 0.0  # user-requested toggle for deviation path
+
+      # Requested deviation length cannot exceed remaining cruise span
+      R_dev = clamp(R_dev_req, 0.0, dR_available) # cannot exceed remaining cruise
+      if Ldebug && R_dev_req > dR_available
+            @warn("Requested deviation length is longer than cruise length, and thus clamped to available distance")
+      end
+
+      # Abort deviation if effective distance collapses to near zero
+      if use_deviation && R_dev <= eps_range
+            if Ldebug
+                  @warn("Deviation skipped: requested deviation distance is negligible.")
+            end
+            use_deviation = false
+      end
+
+      # Get climb rate 
+      # (if deviation climb rate is larger than this, we use climb rate 
+      #  and not the calculated deviation rate)
+      γ_climb = para[iagamV, ipclimb1]
+      γ_abs = abs(γ_climb)
+      eps_gamma = 1.0e-6
+      eps_den = 1.0e-8
+
+      # Compute cruise/deviation mission points
+      function recompute_cruise_point(ipx, γ)
+            # Apply updated flight-path angle to this mission point
+            para[iagamV, ipx] = γ
+
+            # Refresh atmospheric conditions at the current altitude
+            altkm = para[iaalt, ipx] / 1000.0
+            T0, p0, ρ0, a0, μ0 = atmos(altkm, ΔTatmos)
+            pare[iep0, ipx] = p0
+            pare[ieT0, ipx] = T0
+            pare[iea0, ipx] = a0
+            pare[ierho0, ipx] = ρ0
+            pare[iemu0, ipx] = μ0
+
[…]
