### 2025-12-08 fc6026d
added plot profile function
