### 2025-12-10 ca0d6a2
add wake system plot
Update docs

Add tests

add AIC matrix png

minor formatting and the like

docs update
