+++ test/runtests.jl
@@ -8,6 +8,7 @@ include(TASOPT.__TASOPTindices__)
     include("unit_test_structures.jl")
     include("unit_test_loads.jl")
     include("unit_test_aero.jl")
+    include("unit_test_trefftz.jl")
     include("regression_test_size_aircraft.jl")
     include("unit_test_heat_exchanger.jl")
     include("unit_test_ductedfan.jl")
