### 2025-12-10 3dd790a
add wake system plot
Update docs

Add tests

add AIC matrix png

minor formatting and the like

docs update
