+++ test/unit_test_trefftz.jl
@@ -1,0 +1,333 @@
+"""
+Unit tests for Trefftz plane induced drag calculations
+
+This test file covers the refactored functions for Trefftz plane analysis:
+- WakeGeometry.jl: Geometric structs and influence matrix
+- induced_drag.jl: Panel point generation and circulation calculations
+"""
+
+using Test
+using TASOPT
+using LinearAlgebra
+using StaticArrays
+const aero = TASOPT.aerodynamics
+
+@testset "Trefftz calculations" verbose=true begin
+
+    @testset "Wake geometry primitives" begin
+        # Test Point2D creation and basic operations
+        @testset "Point2D" begin
+            p1 = aero.Point2D(1.0, 2.0)
+            p2 = aero.Point2D(4.0, 6.0)
+
+            @test p1[1] == 1.0
+            @test p1[2] == 2.0
+            @test norm(p2 - p1) ≈ 5.0  # 3-4-5 triangle
+        end
+
+        # Test WakeElement construction and properties
+        @testset "WakeElement" begin
+            p1 = aero.Point2D(0.0, 0.0)
+            p2 = aero.Point2D(3.0, 4.0)
+            elem = aero.WakeElement(p1, p2)
+
+            @test elem.length ≈ 5.0
+            @test elem.Δy ≈ 3.0
+            @test elem.Δz ≈ 4.0
+
+            # Control point should be midpoint by default
+            @test elem.control_point[1] ≈ 1.5
+            @test elem.control_point[2] ≈ 2.0
+
+            # Unit normal should be perpendicular and unit length
+            @test norm(elem.unit_normal) ≈ 1.0
+            @test dot(elem.unit_normal, p2 - p1) ≈ 0.0  atol=1e-12# Perpendicular
+            # HA! There's a gotcha here ^ don't try to just compare to zero, need to use atol. 
+            # See `isapprox` documentation for details.
+
+            custom_cp = aero.Point2D(1.0, 1.0)
+            elem_custom = aero.WakeElement(p1, p2; control_point=custom_cp)
+            @test elem_custom.control_point == custom_cp
+
+            @test_throws ArgumentError aero.WakeElement(p1, p1)
+        end
+
+        # Test mirroring function about z-axis
+        @testset "mirror_point" begin
+            p = aero.Point2D(3.0, 5.0)
+            mirrored = aero.mirror_point(p)
+            @test mirrored[1] ≈ -3.0
+            @test mirrored[2] ≈ 5.0  # z unchanged
+        end
+    end
+
+    @testset "Bunching transformations" begin
+        @testset "bunch_transform" begin
+            # Test endpoints
+            @test aero.bunch_transform(0.0, 0.5) ≈ 0.0
+            @test aero.bunch_transform(1.0, 0.5) ≈ 1.0
+
+            # Test no bunching with bunch=0
+            @test aero.bunch_transform(0.3, 0.0) ≈ 0.3
+        end
+
+        @testset "inv_bunch_transform" begin
+            # Test round-trip consistency
+            bunch = 0.5
+            for t in [0.0, 0.1, 0.25, 0.5, 0.75, 0.9, 1.0]
+                t_bunched = aero.bunch_transform(t, bunch)
+                t_recovered = aero.inv_bunch_transform(t_bunched, bunch)
+                @test t_recovered ≈ t rtol=1e-12
+            end
+
+            # Test with different bunch factors
+            for bunch in [0.25, 0.5, 0.75]
+                t = 0.3
+                t_bunched = aero.bunch_transform(t, bunch)
[…]
