+++ docs/src/data_io/structs.md
@@ -54,8 +54,4 @@ TASOPT.materials.StructuralAlloy
 
 aerodynamics.airfoil
 
-aerodynamics.TrefftzPlaneConfig
-
-aerodynamics.SurfaceDiscretization
-
 ```

[…]
