### 2025-12-13 c595d3b
typo and get rid of old _trefftz_analysis
fix broken exmaple link

remove old test

ah forgot to delete some outdated tests from the last commit
