+++ src/aero/induced_drag.jl
@@ -497,118 +497,11 @@ function _build_trefftz_geometry!(wing, htail, po, gammat, gammas,
 end
 
 
-
 """
-    _trefftz_analysis(nsurf, trefftz_config, wing, htail, Sref, bref,
-          po, gammat, gammas, fLo, specifies_CL, CLsurfsp,
-          geom, gw, vc, wc, vnc)
+    compute_induced_drag!(vnc, WS::WakeSystem{Npts, Nelems}, gc, gw, bref, Sref) where {Npts, Nelems}
 
-Trefftz plane routine for the induced drag computation of `nsurf` number of surfaces. Formerly, `trefftz1()`.
-
-!!! details "🔃 Inputs and Outputs"
-    **Inputs:**
-    - `nsurf::Integer`: Number of surfaces (typically wing and horizontal tail).
-    - `trefftz_config::TrefftzPlaneConfig`: Configuration containing panel discretization and analysis parameters.
-    - `wing`: Wing structure (TASOPT.Wing).
-    - `htail`: Horizontal tail structure (TASOPT.Tail).
-    - `Sref::Float64`: Reference wing area.
-    - `bref::Float64`: Reference wing span.
-    - `po::Vector{Float64}`: Root circulation scaling factor for each surface.
-    - `gammat::Vector{Float64}`: Wing lift distribution "taper" ratios for outer sections.
-    - `gammas::Vector{Float64}`: Wing lift distribution "taper" ratios for inner sections.
-    - `fLo::Float64`: Wing root load adjustment factor (currently not implemented).
-    - `specifies_CL::Bool`: Flag for specified lift calculation (scales vorticities to achieve `CLsurfsp`).
-    - `CLsurfsp::Vector{Float64}`: Prescribed surface lift coefficient for each surface.
-    - `geom::TrefftzGeometry`: Geometry arrays (t, y, yp, z, zp, yc, ycp, zc, zcp, gc).
-    - `gw, vc, wc, vnc`: Work arrays for wake circulation and velocities.
-
-    **Outputs:**
-    - `CLsurf::Vector{Float64}`: Lift coefficients for each surface.
-    - `CL::Float64`: Sum of lift coefficients of all surfaces.
-    - `CD::Float64`: Sum of induced drag coefficients of all surfaces.
-    - `spanef::Float64`: Span efficiency of combined surfaces (``= (CL^2 / (π*AR))/CD``).
-
-See [theory above](@ref trefftz) or Sections 2.14.7 and 3.8.1 of the [TASOPT Technical Desc](@ref dreladocs).
+Compute the induced drag and span efficiency from the wake system.
 """
-function _trefftz_analysis(nsurf, trefftz_config::TrefftzPlaneConfig,
-	wing, htail,
-	Sref, bref,
-	po, gammat, gammas, fLo,
-	specifies_CL, CLsurfsp, geom::TrefftzGeometry)
-
-      ifrst = zeros(Int, nsurf)
-      ilast = zeros(Int, nsurf)
-
-      CLsurf= zeros(Float64, nsurf)
-
-    
-    ifrst = [i_first_wing(trefftz_config), i_first_tail(trefftz_config)]
-    ilast = [i_last_wing(trefftz_config), i_last_tail(trefftz_config)]
-
-    _build_trefftz_geometry!(wing, htail, po, gammat, gammas, trefftz_config)
-
-      ii = ilast[nsurf]
-
-      calculate_wake_circulation!(gw, geom.gc, ifrst, ilast, nsurf)
-
-      if(specifies_CL)
[…]
