### 2025-12-13 e254ac3
typo and get rid of old _trefftz_analysis
fix broken exmaple link

remove old test

ah forgot to delete some outdated tests from the last commit
