### 2025-12-16 3038316
Get rid of wc and vc arrays.
