### 2025-12-16 d2ffb87
pretty print add
