### 2025-12-19 59f3399
minor inner consturctor update
