### 2025-12-22 19d4516
Clean up of wakesystem and pretty print
