+++ src/aero/WakeGeometry.jl
@@ -20,9 +20,6 @@ julia> p = Point2D(1.0, 2.0) # Creates point at y=1.0, z=2.0
 """
 const Point2D = SVector{2, Float64} # [y, z] really just a shorthand for SVector
 
-function Base.show(io::IO, p::Point2D)
-    print(io,  @sprintf("[%8.3g, %8.3g]", p[1], p[2]))
-end
 """
     $(TYPEDEF)
 
@@ -69,30 +66,34 @@ end
 
 
 """
-    generate_wake_elements(points::SVector{N, Point2D};
-    control_points::Union{SVector{M, Point2D}, Nothing}=nothing) where {N, M}
+    generate_wake_elements(points::SVector{N, Point2D}) where {N}
 
-Returns an SVector of WakeElements
+Returns an SVector of WakeElements and assumes control points are midpoints of each element
 """
-@inline function generate_wake_elements(points::SVector{N, Point2D};
-    control_points::Union{SVector{M, Point2D}, Nothing}=nothing) where {N, M}
-    if isnothing(control_points)
-        # If no control points are provided, calculate midpoints and set that as the control point
-        return SVector{N-1,WakeElement}(WakeElement(points[i], points[i+1]) for i in 1:N-1)
-    else
-        if M != N - 1
-            throw(ArgumentError("Number of control points must be exactly one less than the number of points."))
-        end
-    return SVector{N-1,WakeElement}(WakeElement(points[i], points[i+1]; 
-                                        control_point=control_points[i]) for i in 1:N-1)
+@inline function generate_wake_elements(points::SVector{N,Point2D}) where {N}
+    # If no control points are provided, calculate midpoints and set that as the control point
+    return SVector{N - 1,WakeElement}(WakeElement(points[i], points[i+1]) for i in 1:N-1)
+end
+
+"""
+    generate_wake_elements(points::SVector{N, Point2D};
+                        control_points::SVector{M, Point2D}) where {N, M}
+
+Returns an SVector of WakeElements using provided points and control points.
+"""
+@inline function generate_wake_elements(points::SVector{N,Point2D};
+    control_points::SVector{M,Point2D}) where {N,M}
+    if M != N - 1
+        throw(ArgumentError("Number of control points must be exactly one less than the number of points."))
     end
+    return SVector{N - 1,WakeElement}(WakeElement(points[i], points[i+1];
+        control_point=control_points[i]) for i in 1:N-1)
 end
 
 #Convenience functions to get element lengths, Δy's and Δz's
 function element_lengths(wake_elements::SVector{N, WakeElement}) where N
     SVector{N, Float64}(wake_elements[i].length for i in 1:N)
 end
-
 
 function element_dys(wake_elements::SVector{N, WakeElement}) where N
     SVector{N, Float64}(wake_elements[i].Δy for i in 1:N)
@@ -101,10 +102,6 @@ end  # function element_dys
 function element_dzs(wake_elements::SVector{N, WakeElement}) where N
     SVector{N, Float64}(wake_elements[i].Δz for i in 1:N)
 end  # function element_dzs
-
-function Base.show(io::IO, el::WakeElement)
-    print(io, "WakeElement($(el.p1) -- $(el.p2))")
-end
 
 """
     $(TYPEDEF)
@@ -137,7 +134,11 @@ struct WakeSystem{NP, NE}
         if NE != NP - 1 || AIC_size[1] != (AIC_size[2] - 1) || AIC_size[1] != NE
[…]
