### 2025-12-22 52daf3d
Clean up of wakesystem and pretty print
