### 2025-12-24 7b75b98
extract weight initialization to separate file
