+++ src/sizing/size_aircraft.jl
@@ -154,157 +154,25 @@ function _size_aircraft!(ac; itermax=35,
     # Reset engine values for heat exchangers
     resetHXs(pare)
 
-    # -------------------------------------------------------    
+    # -------------------------------------------------------
     ## Initial guess section [Section 3.2 of TASOPT docs]
     # -------------------------------------------------------
-    # Allow first iteration
-    if (initwgt == 0)
+    if initwgt == 0
+        # Initialize weights, geometry, and parameters for first sizing run
+        initialize_sizing_loop!(ac)
 
-        # Initial weight estimates
-        Whtail = 0.05 * Wpay / parg[igsigfac]
-        Wvtail = Whtail
-        Wwing = 0.5 * Wpay / parg[igsigfac]
-        Wstrut = 0.0
-        Wftank = 0.0
-        Weng = 0.0 * Wpay
-        feng = 0.0
-
-        dxWhtail = 0.0
-        dxWvtail = 0.0
-
-        # Wing panel weights and moments
-        ip = ipcruise1
-        W = 5.0 * Wpay
-        S = W / (0.5 * pare[ierho0, ip] * pare[ieu0, ip]^2 * para[iaCL, ip])
-        b = sqrt(S * wing.layout.AR)
-        bs = b * wing.layout.ηs
-        Winn = 0.15 * Wpay / parg[igsigfac]
-        Wout = 0.05 * Wpay / parg[igsigfac]
-        dyWinn = Winn * 0.30 * (0.5 * (bs - wing.layout.root_span))
-        dyWout = Wout * 0.25 * (0.5 * (b - bs))
-
-        # Assign weights to components
-        htail.weight = Whtail
-        vtail.weight = Wvtail
-        wing.weight = Wwing
-        wing.strut.weight = Wstrut
-        parg[igWeng] = Weng
-        wing.inboard.weight = Winn
-        wing.outboard.weight = Wout
-        parg[igWftank] = Wftank
-        htail.dxW = dxWhtail
-        vtail.dxW = dxWvtail
-        wing.inboard.dyW = dyWinn
-        wing.outboard.dyW = dyWout
-
-        # Wing centroid x-offset from wingbox
-        calculate_centroid_offset!(wing, b=b, bs=bs)
-
-        # Tail area centroid locations
-        htail.layout.x, vtail.layout.x = xhbox, xvbox
-
-        # Center wingbox chord extent for fuselage weight calculations
+        # Extract working variables set by initialization
+        Sh = htail.layout.S
+        Sv = vtail.layout.S
+        bv = sqrt(Sv * vtail.layout.AR)
         cbox = 0.0
 
-        # Nacelle, fan duct, core, cowl lengths
-        parg[iglnace] = 0.5 * S / b
-
-        # Nacelle Awet/S
-        fSnace = 0.2
-        parg[igfSnace] = fSnace
-
-        # Initial fuel fraction estimate from Breguet Range Equation
-        LoD = 18.0
-        TSFC = 1.0 / 7000.0
-        V = pare[ieu0, ipcruise1]
-        ffburn = min((1.0 - exp(-Rangetot * TSFC / (V * LoD))), 0.8 / (1.0 + freserve))
-
-        # Mission-point fuel fractions
-        ffuelb = ffburn * (1.0 + freserve)  # Start of climb
-        ffuelc = ffburn * (0.90 + freserve)  # Start of cruise
[…]
