+++ src/sizing/initialize_weights.jl
@@ -1,0 +1,212 @@
+"""
+    initialize_sizing_loop!(ac)
+
+Initializes weights, geometry estimates, and parameters for the first iteration
+of the aircraft sizing loop. Sets all state directly in the aircraft structs and
+parameter arrays.
+
+This function computes initial estimates using:
+- Payload-based weight fractions for components
+- Breguet range equation for fuel fraction
+- Simple geometric relationships for wing/tail sizing
+
+All values are approximate and will be refined during iteration.
+"""
+function initialize_sizing_loop!(ac)
+    # Unpack aircraft components and parameter arrays
+    imission = 1
+    parg, parm, para, pare, _, fuse, _, wing, htail, vtail, _, landing_gear = unpack_ac(ac, imission)
+
+    # Extract design parameters
+    Wpay = parm[imWpay]
+    Wpaymax = parg[igWpaymax]
+    Rangetot = parm[imRange]
+    freserve = parg[igfreserve]
+    neng = parg[igneng]
+
+    # Extract layout parameters
+    xhbox = htail.layout.box_x
+    xvbox = vtail.layout.box_x
+
+    # Landing gear weight fractions
+    flgnose = landing_gear.nose_gear.overall_mass_fraction
+    flgmain = landing_gear.main_gear.overall_mass_fraction
+
+    # Dynamic pressure at never-exceed speed
+    Vne = parg[igVne]
+    qne = 0.5 * ρSL * Vne^2
+
+    # Structural weight factor
+    sigfac = parg[igsigfac]
+
+    # ===== Initial weight estimates =====
+    Whtail = 0.05 * Wpay / sigfac
+    Wvtail = Whtail
+    Wwing = 0.5 * Wpay / sigfac
+    Wstrut = 0.0
+    Wftank = 0.0
+    Weng = 0.0
+
+    # ===== Wing geometry estimates =====
+    ip = ipcruise1
+    W_estimate = 5.0 * Wpay  # Initial weight estimate
+    S = W_estimate / (0.5 * pare[ierho0, ip] * pare[ieu0, ip]^2 * para[iaCL, ip])
+    b = sqrt(S * wing.layout.AR)
+    bs = b * wing.layout.ηs
+
+    # Wing panel weights
+    Winn = 0.15 * Wpay / sigfac
+    Wout = 0.05 * Wpay / sigfac
+    dyWinn = Winn * 0.30 * (0.5 * (bs - wing.layout.root_span))
+    dyWout = Wout * 0.25 * (0.5 * (b - bs))
+
+    # ===== Store initial weights in components =====
+    htail.weight = Whtail
+    vtail.weight = Wvtail
+    wing.weight = Wwing
+    wing.strut.weight = Wstrut
+    wing.inboard.weight = Winn
+    wing.outboard.weight = Wout
+
+    parg[igWeng] = Weng
+    parg[igWftank] = Wftank
+
+    # Weight moments
+    htail.dxW = 0.0
+    vtail.dxW = 0.0
+    wing.inboard.dyW = dyWinn
+    wing.outboard.dyW = dyWout
+
+    # Wing centroid x-offset from wingbox
+    calculate_centroid_offset!(wing, b=b, bs=bs)
+
+    # Tail area centroid locations
+    htail.layout.x = xhbox
+    vtail.layout.x = xvbox
+
[…]
