### 2025-12-24 b396398
remove dead code
fix stray =
