### 2025-12-24 42ce877
clean up weight initialization
minor update

minor fix
