### 2025-12-25 76022de
refactor fuel tank initialization
