### 2025-12-25 c22aa5e
remove old unused argument
