### 2025-12-26 6389664
Add flight condition struct
include flight conditions and atmos state into TASOPT

udpate ac/ flight condition load order
