### 2025-12-26 0e16d59
add flight conditions to aircraft
