+++ src/data_structs/aircraft.jl
@@ -30,6 +30,7 @@ Overloads Base.summary to print a summary of the `aircraft` model.
 - `htail::Tail` : Horizontal tail data and parameters
 - `vtail::Tail` : Vertical tail data and parameters
 - `engine::Engine` : Engine models, data, and parameters
+- `flight_conditions::Matrix{FlightCondition}` : Flight conditions at each mission point [iptotal × n_missions]
 
 The indices for accessing specific data in the `par` arrays are defined in `/src/data_structs/index.inc`. 
 Refer to the sample input file (`/example/defaults/default_input.toml` and `read_input.jl`) for usage.
@@ -55,6 +56,11 @@ Refer to the docs for a summary of the main `struct`s.
     vtail::Tail = Tail()
     engine::Engine = Engine()
     landing_gear::LandingGear = LandingGear()
+
+    # Flight conditions at each mission point [iptotal × n_missions]
+    # Replaces para[iaalt/iaMach/...] and pare[ieT0/iep0/...] indices
+    flight_conditions::Matrix{FlightCondition}
+
     #TODO: update DOCSTRING for ANY NEW fields/sub-structures
     wake_system::WS
 end
@@ -74,7 +80,8 @@ function aircraft(
     htail::Tail,
     vtail::Tail,
     engine::Engine,
-    landing_gear::LandingGear
+    landing_gear::LandingGear,
+    flight_conditions::Matrix{FlightCondition}
 )
     # Create placeholder WakeSystem with correct size for type stability
     # This will be rebuilt with actual geometry in induced_drag!
@@ -87,6 +94,7 @@ function aircraft(
         fuselage, fuse_tank,
         wing, htail, vtail,
         engine, landing_gear,
+        flight_conditions,
         wake_system
     )
 end
