### 2025-12-26 d96f993
update the ambient conditions function
