### 2025-12-26 3013cae
initialize flight conditions
