### 2025-12-26 ad1a167
update output plots to use flight conditions.
