### 2025-12-26 be1b0a9
add unit tests for flight conditions
